Update externalapi/monitoring/v1 to v0.80.0
