+++ pkg/externalapi/monitoring/v1/types.go
@@ -12,369 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// From https://github.com/prometheus-operator/prometheus-operator/blob/56cc9eea5f8bffedbc4a77ae08555dd5f510ed76/pkg/apis/monitoring/v1/types.go.
+
 package v1
 
 import (
 	v1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
 const (
 	Version = "v1"
-
-	PrometheusesKind  = "Prometheus"
-	PrometheusName    = "prometheuses"
-	PrometheusKindKey = "prometheus"
-
-	AlertmanagersKind   = "Alertmanager"
-	AlertmanagerName    = "alertmanagers"
-	AlertManagerKindKey = "alertmanager"
-
-	ServiceMonitorsKind   = "ServiceMonitor"
-	ServiceMonitorName    = "servicemonitors"
-	ServiceMonitorKindKey = "servicemonitor"
-
-	PodMonitorsKind   = "PodMonitor"
-	PodMonitorName    = "podmonitors"
-	PodMonitorKindKey = "podmonitor"
-
-	PrometheusRuleKind    = "PrometheusRule"
-	PrometheusRuleName    = "prometheusrules"
-	PrometheusRuleKindKey = "prometheusrule"
-
-	ProbesKind   = "Probe"
-	ProbeName    = "probes"
-	ProbeKindKey = "probe"
 )
-
-var resourceToKind = map[string]string{
-	PrometheusName:     PrometheusesKind,
-	AlertmanagerName:   AlertmanagersKind,
-	ServiceMonitorName: ServiceMonitorsKind,
-	PodMonitorName:     PodMonitorsKind,
-	PrometheusRuleName: PrometheusRuleKind,
-	ProbeName:          ProbesKind,
-}
-
-// CommonPrometheusFields are the options available to both the Prometheus server and agent.
-// +k8s:deepcopy-gen=true
-type CommonPrometheusFields struct {
-	// PodMetadata configures Labels and Annotations which are propagated to the prometheus pods.
-	PodMetadata *EmbeddedObjectMetadata `json:"podMetadata,omitempty"`
-	// ServiceMonitors to be selected for target discovery. *Deprecated:* if
-	// neither this nor podMonitorSelector are specified, configuration is
-	// unmanaged.
-	ServiceMonitorSelector *metav1.LabelSelector `json:"serviceMonitorSelector,omitempty"`
-	// Namespace's labels to match for ServiceMonitor discovery. If nil, only
-	// check own namespace.
-	ServiceMonitorNamespaceSelector *metav1.LabelSelector `json:"serviceMonitorNamespaceSelector,omitempty"`
-	// *Experimental* PodMonitors to be selected for target discovery.
-	// *Deprecated:* if neither this nor serviceMonitorSelector are specified,
-	// configuration is unmanaged.
-	PodMonitorSelector *metav1.LabelSelector `json:"podMonitorSelector,omitempty"`
-	// Namespace's labels to match for PodMonitor discovery. If nil, only
-	// check own namespace.
-	PodMonitorNamespaceSelector *metav1.LabelSelector `json:"podMonitorNamespaceSelector,omitempty"`
-	// *Experimental* Probes to be selected for target discovery.
-	ProbeSelector *metav1.LabelSelector `json:"probeSelector,omitempty"`
-	// *Experimental* Namespaces to be selected for Probe discovery. If nil, only check own namespace.
-	ProbeNamespaceSelector *metav1.LabelSelector `json:"probeNamespaceSelector,omitempty"`
-	// Version of Prometheus to be deployed.
-	Version string `json:"version,omitempty"`
-	// When a Prometheus deployment is paused, no actions except for deletion
-	// will be performed on the underlying objects.
-	Paused bool `json:"paused,omitempty"`
-	// Image if specified has precedence over baseImage, tag and sha
-	// combinations. Specifying the version is still necessary to ensure the
-	// Prometheus Operator knows what version of Prometheus is being
-	// configured.
-	Image *string `json:"image,omitempty"`
-	// An optional list of references to secrets in the same namespace
-	// to use for pulling prometheus and alertmanager images from registries
-	// see http://kubernetes.io/docs/user-guide/images#specifying-imagepullsecrets-on-a-pod
-	ImagePullSecrets []v1.LocalObjectReference `json:"imagePullSecrets,omitempty"`
-	// Number of replicas of each shard to deploy for a Prometheus deployment.
-	// Number of replicas multiplied by shards is the total number of Pods
-	// created.
-	Replicas *int32 `json:"replicas,omitempty"`
-	// EXPERIMENTAL: Number of shards to distribute targets onto. Number of
-	// replicas multiplied by shards is the total number of Pods created. Note
-	// that scaling down shards will not reshard data onto remaining instances,
-	// it must be manually moved. Increasing shards will not reshard data
-	// either but it will continue to be available from the same instances. To
-	// query globally use Thanos sidecar and Thanos querier or remote write
-	// data to a central location. Sharding is done on the content of the
-	// `__address__` target meta-label.
-	Shards *int32 `json:"shards,omitempty"`
-	// Name of Prometheus external label used to denote replica name.
-	// Defaults to the value of `prometheus_replica`. External label will
-	// _not_ be added when value is set to empty string (`""`).
-	ReplicaExternalLabelName *string `json:"replicaExternalLabelName,omitempty"`
-	// Name of Prometheus external label used to denote Prometheus instance
-	// name. Defaults to the value of `prometheus`. External label will
-	// _not_ be added when value is set to empty string (`""`).
-	PrometheusExternalLabelName *string `json:"prometheusExternalLabelName,omitempty"`
-	// Log level for Prometheus to be configured with.
-	// +kubebuilder:validation:Enum="";debug;info;warn;error
-	LogLevel string `json:"logLevel,omitempty"`
-	// Log format for Prometheus to be configured with.
-	// +kubebuilder:validation:Enum="";logfmt;json
-	LogFormat string `json:"logFormat,omitempty"`
-	// Interval between consecutive scrapes. Default: `30s`
-	// +kubebuilder:default:="30s"
-	ScrapeInterval Duration `json:"scrapeInterval,omitempty"`
-	// Number of seconds to wait for target to respond before erroring.
-	ScrapeTimeout Duration `json:"scrapeTimeout,omitempty"`
-	// The labels to add to any time series or alerts when communicating with
-	// external systems (federation, remote storage, Alertmanager).
-	ExternalLabels map[string]string `json:"externalLabels,omitempty"`
-	// Enable Prometheus to be used as a receiver for the Prometheus remote write protocol. Defaults to the value of `false`.
-	// WARNING: This is not considered an efficient way of ingesting samples.
-	// Use it with caution for specific low-volume use cases.
-	// It is not suitable for replacing the ingestion via scraping and turning
-	// Prometheus into a push-based metrics collection system.
-	// For more information see https://prometheus.io/docs/prometheus/latest/querying/api/#remote-write-receiver
-	// Only valid in Prometheus versions 2.33.0 and newer.
-	EnableRemoteWriteReceiver bool `json:"enableRemoteWriteReceiver,omitempty"`
-	// Enable access to Prometheus disabled features. By default, no features are enabled.
-	// Enabling disabled features is entirely outside the scope of what the maintainers will
-	// support and by doing so, you accept that this behaviour may break at any
-	// time without notice.
-	// For more information see https://prometheus.io/docs/prometheus/latest/disabled_features/
-	EnableFeatures []string `json:"enableFeatures,omitempty"`
-	// The external URL the Prometheus instances will be available under. This is
-	// necessary to generate correct URLs. This is necessary if Prometheus is not
-	// served from root of a DNS name.
-	ExternalURL string `json:"externalUrl,omitempty"`
-	// The route prefix Prometheus registers HTTP handlers for. This is useful,
-	// if using ExternalURL and a proxy is rewriting HTTP routes of a request,
-	// and the actual ExternalURL is still true, but the server serves requests
-	// under a different route prefix. For example for use with `kubectl proxy`.
-	RoutePrefix string `json:"routePrefix,omitempty"`
-	// Storage spec to specify how storage shall be used.
-	Storage *StorageSpec `json:"storage,omitempty"`
-	// Volumes allows configuration of additional volumes on the output StatefulSet definition. Volumes specified will
-	// be appended to other volumes that are generated as a result of StorageSpec objects.
-	Volumes []v1.Volume `json:"volumes,omitempty"`
-	// VolumeMounts allows configuration of additional VolumeMounts on the output StatefulSet definition.
-	// VolumeMounts specified will be appended to other VolumeMounts in the prometheus container,
-	// that are generated as a result of StorageSpec objects.
-	VolumeMounts []v1.VolumeMount `json:"volumeMounts,omitempty"`
-	// Defines the web command line flags when starting Prometheus.
-	Web *PrometheusWebSpec `json:"web,omitempty"`
-	// Define resources requests and limits for single Pods.
-	Resources v1.ResourceRequirements `json:"resources,omitempty"`
-	// Define which Nodes the Pods are scheduled on.
-	NodeSelector map[string]string `json:"nodeSelector,omitempty"`
-	// ServiceAccountName is the name of the ServiceAccount to use to run the
-	// Prometheus Pods.
-	ServiceAccountName string `json:"serviceAccountName,omitempty"`
-	// Secrets is a list of Secrets in the same namespace as the Prometheus
-	// object, which shall be mounted into the Prometheus Pods.
-	// Each Secret is added to the StatefulSet definition as a volume named `secret-<secret-name>`.
-	// The Secrets are mounted into /etc/prometheus/secrets/<secret-name> in the 'prometheus' container.
-	Secrets []string `json:"secrets,omitempty"`
-	// ConfigMaps is a list of ConfigMaps in the same namespace as the Prometheus
-	// object, which shall be mounted into the Prometheus Pods.
-	// Each ConfigMap is added to the StatefulSet definition as a volume named `configmap-<configmap-name>`.
-	// The ConfigMaps are mounted into /etc/prometheus/configmaps/<configmap-name> in the 'prometheus' container.
-	ConfigMaps []string `json:"configMaps,omitempty"`
-	// If specified, the pod's scheduling constraints.
-	Affinity *v1.Affinity `json:"affinity,omitempty"`
-	// If specified, the pod's tolerations.
-	Tolerations []v1.Toleration `json:"tolerations,omitempty"`
-	// If specified, the pod's topology spread constraints.
-	TopologySpreadConstraints []v1.TopologySpreadConstraint `json:"topologySpreadConstraints,omitempty"`
-	// remoteWrite is the list of remote write configurations.
-	RemoteWrite []RemoteWriteSpec `json:"remoteWrite,omitempty"`
-	// SecurityContext holds pod-level security attributes and common container settings.
-	// This defaults to the default PodSecurityContext.
-	SecurityContext *v1.PodSecurityContext `json:"securityContext,omitempty"`
-	// ListenLocal makes the Prometheus server listen on loopback, so that it
-	// does not bind against the Pod IP.
-	ListenLocal bool `json:"listenLocal,omitempty"`
-	// Containers allows injecting additional containers or modifying operator
-	// generated containers. This can be used to allow adding an authentication
-	// proxy to a Prometheus pod or to change the behavior of an operator
-	// generated container. Containers described here modify an operator
-	// generated container if they share the same name and modifications are
-	// done via a strategic merge patch. The current container names are:
-	// `prometheus`, `config-reloader`, and `thanos-sidecar`. Overriding
-	// containers is entirely outside the scope of what the maintainers will
-	// support and by doing so, you accept that this behaviour may break at any
-	// time without notice.
-	Containers []v1.Container `json:"containers,omitempty"`
-	// InitContainers allows adding initContainers to the pod definition. Those can be used to e.g.
-	// fetch secrets for injection into the Prometheus configuration from external sources. Any errors
-	// during the execution of an initContainer will lead to a restart of the Pod. More info: https://kubernetes.io/docs/concepts/workloads/pods/init-containers/
-	// InitContainers described here modify an operator
-	// generated init containers if they share the same name and modifications are
-	// done via a strategic merge patch. The current init container name is:
-	// `init-config-reloader`. Overriding init containers is entirely outside the
-	// scope of what the maintainers will support and by doing so, you accept that
-	// this behaviour may break at any time without notice.
-	InitContainers []v1.Container `json:"initContainers,omitempty"`
-	// AdditionalScrapeConfigs allows specifying a key of a Secret containing
-	// additional Prometheus scrape configurations. Scrape configurations
-	// specified are appended to the configurations generated by the Prometheus
-	// Operator. Job configurations specified must have the form as specified
-	// in the official Prometheus documentation:
-	// https://prometheus.io/docs/prometheus/latest/configuration/configuration/#scrape_config.
-	// As scrape configs are appended, the user is responsible to make sure it
-	// is valid. Note that using this feature may expose the possibility to
-	// break upgrades of Prometheus. It is advised to review Prometheus release
-	// notes to ensure that no incompatible scrape configs are going to break
-	// Prometheus after the upgrade.
-	AdditionalScrapeConfigs *v1.SecretKeySelector `json:"additionalScrapeConfigs,omitempty"`
-	// APIServerConfig allows specifying a host and auth methods to access apiserver.
-	// If left empty, Prometheus is assumed to run inside of the cluster
-	// and will discover API servers automatically and use the pod's CA certificate
-	// and bearer token file at /var/run/secrets/kubernetes.io/serviceaccount/.
-	APIServerConfig *APIServerConfig `json:"apiserverConfig,omitempty"`
-	// Priority class assigned to the Pods
-	PriorityClassName string `json:"priorityClassName,omitempty"`
-	// Port name used for the pods and governing service.
-	// This defaults to web
-	PortName string `json:"portName,omitempty"`
-	// ArbitraryFSAccessThroughSMs configures whether configuration
-	// based on a service monitor can access arbitrary files on the file system
-	// of the Prometheus container e.g. bearer token files.
-	ArbitraryFSAccessThroughSMs ArbitraryFSAccessThroughSMsConfig `json:"arbitraryFSAccessThroughSMs,omitempty"`
-	// When true, Prometheus resolves label conflicts by renaming the labels in
-	// the scraped data to "exported_<label value>" for all targets created
-	// from service and pod monitors.
-	// Otherwise the HonorLabels field of the service or pod monitor applies.
-	OverrideHonorLabels bool `json:"overrideHonorLabels,omitempty"`
-	// When true, Prometheus ignores the timestamps for all the targets created
-	// from service and pod monitors.
-	// Otherwise the HonorTimestamps field of the service or pod monitor applies.
-	OverrideHonorTimestamps bool `json:"overrideHonorTimestamps,omitempty"`
-	// IgnoreNamespaceSelectors if set to true will ignore NamespaceSelector
-	// settings from all PodMonitor, ServiceMonitor and Probe objects. They will
-	// only discover endpoints within the namespace of the PodMonitor,
-	// ServiceMonitor and Probe objects.
-	// Defaults to false.
-	IgnoreNamespaceSelectors bool `json:"ignoreNamespaceSelectors,omitempty"`
-	// EnforcedNamespaceLabel If set, a label will be added to
-	//
-	// 1. all user-metrics (created by `ServiceMonitor`, `PodMonitor` and `Probe` objects) and
-	// 2. in all `PrometheusRule` objects (except the ones excluded in `prometheusRulesExcludedFromEnforce`) to
-	//    * alerting & recording rules and
-	//    * the metrics used in their expressions (`expr`).
-	//
-	// Label name is this field's value.
-	// Label value is the namespace of the created object (mentioned above).
-	EnforcedNamespaceLabel string `json:"enforcedNamespaceLabel,omitempty"`
-	// EnforcedSampleLimit defines global limit on number of scraped samples
-	// that will be accepted. This overrides any SampleLimit set per
-	// ServiceMonitor or/and PodMonitor. It is meant to be used by admins to
-	// enforce the SampleLimit to keep overall number of samples/series under
-	// the desired limit.
-	// Note that if SampleLimit is lower that value will be taken instead.
-	EnforcedSampleLimit *uint64 `json:"enforcedSampleLimit,omitempty"`
-	// EnforcedTargetLimit defines a global limit on the number of scraped
-	// targets.  This overrides any TargetLimit set per ServiceMonitor or/and
-	// PodMonitor.  It is meant to be used by admins to enforce the TargetLimit
-	// to keep the overall number of targets under the desired limit.
-	// Note that if TargetLimit is lower, that value will be taken instead,
-	// except if either value is zero, in which case the non-zero value will be
-	// used.  If both values are zero, no limit is enforced.
-	EnforcedTargetLimit *uint64 `json:"enforcedTargetLimit,omitempty"`
-	// Per-scrape limit on number of labels that will be accepted for a sample. If
-	// more than this number of labels are present post metric-relabeling, the
-	// entire scrape will be treated as failed. 0 means no limit.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	EnforcedLabelLimit *uint64 `json:"enforcedLabelLimit,omitempty"`
-	// Per-scrape limit on length of labels name that will be accepted for a sample.
-	// If a label name is longer than this number post metric-relabeling, the entire
-	// scrape will be treated as failed. 0 means no limit.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	EnforcedLabelNameLengthLimit *uint64 `json:"enforcedLabelNameLengthLimit,omitempty"`
-	// Per-scrape limit on length of labels value that will be accepted for a sample.
-	// If a label value is longer than this number post metric-relabeling, the
-	// entire scrape will be treated as failed. 0 means no limit.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	EnforcedLabelValueLengthLimit *uint64 `json:"enforcedLabelValueLengthLimit,omitempty"`
-	// EnforcedBodySizeLimit defines the maximum size of uncompressed response body
-	// that will be accepted by Prometheus. Targets responding with a body larger than this many bytes
-	// will cause the scrape to fail. Example: 100MB.
-	// If defined, the limit will apply to all service/pod monitors and probes.
-	// This is an experimental feature, this behaviour could
-	// change or be removed in the future.
-	// Only valid in Prometheus versions 2.28.0 and newer.
-	EnforcedBodySizeLimit ByteSize `json:"enforcedBodySizeLimit,omitempty"`
-	// Minimum number of seconds for which a newly created pod should be ready
-	// without any of its container crashing for it to be considered available.
-	// Defaults to 0 (pod will be considered available as soon as it is ready)
-	// This is an alpha field and requires enabling StatefulSetMinReadySeconds feature gate.
-	// +optional
-	MinReadySeconds *uint32 `json:"minReadySeconds,omitempty"`
-	// Pods' hostAliases configuration
-	// +listType=map
-	// +listMapKey=ip
-	HostAliases []HostAlias `json:"hostAliases,omitempty"`
-	// AdditionalArgs allows setting additional arguments for the Prometheus container.
-	// It is intended for e.g. activating hidden flags which are not supported by
-	// the dedicated configuration options yet. The arguments are passed as-is to the
-	// Prometheus container which may cause issues if they are invalid or not supported
-	// by the given Prometheus version.
-	// In case of an argument conflict (e.g. an argument which is already set by the
-	// operator itself) or when providing an invalid argument the reconciliation will
-	// fail and an error will be logged.
-	AdditionalArgs []Argument `json:"additionalArgs,omitempty"`
-	// Enable compression of the write-ahead log using Snappy. This flag is
-	// only available in versions of Prometheus >= 2.11.0.
-	WALCompression *bool `json:"walCompression,omitempty"`
-	// List of references to PodMonitor, ServiceMonitor, Probe and PrometheusRule objects
-	// to be excluded from enforcing a namespace label of origin.
-	// Applies only if enforcedNamespaceLabel set to true.
-	ExcludedFromEnforcement []ObjectReference `json:"excludedFromEnforcement,omitempty"`
-	// Use the host's network namespace if true.
-	// Make sure to understand the security implications if you want to enable it.
-	// When hostNetwork is enabled, this will set dnsPolicy to ClusterFirstWithHostNet automatically.
-	HostNetwork bool `json:"hostNetwork,omitempty"`
-}
-
-// +kubebuilder:object:root=true
-// +kubebuilder:storageversion
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +genclient
-// +kubebuilder:resource:categories="prometheus-operator",shortName="prom"
-// +kubebuilder:printcolumn:name="Version",type="string",JSONPath=".spec.version",description="The version of Prometheus"
-// +kubebuilder:printcolumn:name="Desired",type="integer",JSONPath=".spec.replicas",description="The number of desired replicas"
-// +kubebuilder:printcolumn:name="Ready",type="integer",JSONPath=".status.availableReplicas",description="The number of ready replicas"
-// +kubebuilder:printcolumn:name="Reconciled",type="string",JSONPath=".status.conditions[?(@.type == 'Reconciled')].status"
-// +kubebuilder:printcolumn:name="Available",type="string",JSONPath=".status.conditions[?(@.type == 'Available')].status"
-// +kubebuilder:printcolumn:name="Age",type="date",JSONPath=".metadata.creationTimestamp"
-// +kubebuilder:printcolumn:name="Paused",type="boolean",JSONPath=".status.paused",description="Whether the resource reconciliation is paused or not",priority=1
-// +kubebuilder:subresource:status
-
-// Prometheus defines a Prometheus deployment.
-type Prometheus struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of the desired behavior of the Prometheus cluster. More info:
-	// https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
-	Spec PrometheusSpec `json:"spec"`
-	// Most recent observed status of the Prometheus cluster. Read-only.
-	// More info:
-	// https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
-	Status PrometheusStatus `json:"status,omitempty"`
-}
-
-// +kubebuilder:object:root=true
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
-// PrometheusList is a list of Prometheuses.
-type PrometheusList struct {
-	metav1.TypeMeta `json:",inline"`
-	// Standard list metadata
-	// More info: https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#metadata
-	metav1.ListMeta `json:"metadata,omitempty"`
-	// List of Prometheuses
-	Items []*Prometheus `json:"items"`
-}
 
 // ByteSize is a valid memory size type based on powers-of-2, so 1KB is 1024B.
 // Supported units: B, KB, KiB, MB, MiB, GB, GiB, TB, TiB, PB, PiB, EB, EiB Ex: `512MB`.
@@ -386,6 +37,14 @@ type ByteSize string
 // Examples: `30s`, `1m`, `1h20m15s`, `15d`
 // +kubebuilder:validation:Pattern:="^(0|(([0-9]+)y)?(([0-9]+)w)?(([0-9]+)d)?(([0-9]+)h)?(([0-9]+)m)?(([0-9]+)s)?(([0-9]+)ms)?)$"
 type Duration string
+
+// NonEmptyDuration is a valid time duration that can be parsed by Prometheus model.ParseDuration() function.
+// Compared to Duration,  NonEmptyDuration enforces a minimum length of 1.
+// Supported units: y, w, d, h, m, s, ms
+// Examples: `30s`, `1m`, `1h20m15s`, `15d`
+// +kubebuilder:validation:Pattern:="^(0|(([0-9]+)y)?(([0-9]+)w)?(([0-9]+)d)?(([0-9]+)h)?(([0-9]+)m)?(([0-9]+)s)?(([0-9]+)ms)?)$"
+// +kubebuilder:validation:MinLength=1
+type NonEmptyDuration string
 
 // GoDuration is a valid time duration that can be parsed by Go's time.ParseDuration() function.
 // Supported units: h, m, s, ms
@@ -404,141 +63,41 @@ type HostAlias struct {
 	Hostnames []string `json:"hostnames"`
 }
 
-// PrometheusSpec is a specification of the desired behavior of the Prometheus cluster. More info:
-// https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
-type PrometheusSpec struct {
-	CommonPrometheusFields `json:",inline"`
-	// Base image to use for a Prometheus deployment.
-	// Deprecated: use 'image' instead
-	BaseImage string `json:"baseImage,omitempty"`
-	// Tag of Prometheus container image to be deployed. Defaults to the value of `version`.
-	// Version is ignored if Tag is set.
-	// Deprecated: use 'image' instead.  The image tag can be specified
-	// as part of the image URL.
-	Tag string `json:"tag,omitempty"`
-	// SHA of Prometheus container image to be deployed. Defaults to the value of `version`.
-	// Similar to a tag, but the SHA explicitly deploys an immutable container image.
-	// Version and Tag are ignored if SHA is set.
-	// Deprecated: use 'image' instead.  The image digest can be specified
-	// as part of the image URL.
-	SHA string `json:"sha,omitempty"`
-	// Time duration Prometheus shall retain data for. Default is '24h' if
-	// retentionSize is not set, and must match the regular expression `[0-9]+(ms|s|m|h|d|w|y)`
-	// (milliseconds seconds minutes hours days weeks years).
-	Retention Duration `json:"retention,omitempty"`
-	// Maximum amount of disk space used by blocks.
-	RetentionSize ByteSize `json:"retentionSize,omitempty"`
-	// Disable prometheus compaction.
-	DisableCompaction bool `json:"disableCompaction,omitempty"`
-	// /--rules.*/ command-line arguments.
-	Rules Rules `json:"rules,omitempty"`
-	// PrometheusRulesExcludedFromEnforce - list of prometheus rules to be excluded from enforcing
-	// of adding namespace labels. Works only if enforcedNamespaceLabel set to true.
-	// Make sure both ruleNamespace and ruleName are set for each pair.
-	// Deprecated: use excludedFromEnforcement instead.
-	PrometheusRulesExcludedFromEnforce []PrometheusRuleExcludeConfig `json:"prometheusRulesExcludedFromEnforce,omitempty"`
-	// QuerySpec defines the query command line flags when starting Prometheus.
-	Query *QuerySpec `json:"query,omitempty"`
-	// A selector to select which PrometheusRules to mount for loading alerting/recording
-	// rules from. Until (excluding) Prometheus Operator v0.24.0 Prometheus
-	// Operator will migrate any legacy rule ConfigMaps to PrometheusRule custom
-	// resources selected by RuleSelector. Make sure it does not match any config
-	// maps that you do not want to be migrated.
-	RuleSelector *metav1.LabelSelector `json:"ruleSelector,omitempty"`
-	// Namespaces to be selected for PrometheusRules discovery. If unspecified, only
-	// the same namespace as the Prometheus object is in is used.
-	RuleNamespaceSelector *metav1.LabelSelector `json:"ruleNamespaceSelector,omitempty"`
-	// Define details regarding alerting.
-	Alerting *AlertingSpec `json:"alerting,omitempty"`
-	// remoteRead is the list of remote read configurations.
-	RemoteRead []RemoteReadSpec `json:"remoteRead,omitempty"`
-	// AdditionalAlertRelabelConfigs allows specifying a key of a Secret containing
-	// additional Prometheus alert relabel configurations. Alert relabel configurations
-	// specified are appended to the configurations generated by the Prometheus
-	// Operator. Alert relabel configurations specified must have the form as specified
-	// in the official Prometheus documentation:
-	// https://prometheus.io/docs/prometheus/latest/configuration/configuration/#alert_relabel_configs.
-	// As alert relabel configs are appended, the user is responsible to make sure it
-	// is valid. Note that using this feature may expose the possibility to
-	// break upgrades of Prometheus. It is advised to review Prometheus release
-	// notes to ensure that no incompatible alert relabel configs are going to break
-	// Prometheus after the upgrade.
-	AdditionalAlertRelabelConfigs *v1.SecretKeySelector `json:"additionalAlertRelabelConfigs,omitempty"`
-	// AdditionalAlertManagerConfigs allows specifying a key of a Secret containing
-	// additional Prometheus AlertManager configurations. AlertManager configurations
-	// specified are appended to the configurations generated by the Prometheus
-	// Operator. Job configurations specified must have the form as specified
-	// in the official Prometheus documentation:
-	// https://prometheus.io/docs/prometheus/latest/configuration/configuration/#alertmanager_config.
-	// As AlertManager configs are appended, the user is responsible to make sure it
-	// is valid. Note that using this feature may expose the possibility to
-	// break upgrades of Prometheus. It is advised to review Prometheus release
-	// notes to ensure that no incompatible AlertManager configs are going to break
-	// Prometheus after the upgrade.
-	AdditionalAlertManagerConfigs *v1.SecretKeySelector `json:"additionalAlertManagerConfigs,omitempty"`
-	// Thanos configuration allows configuring various aspects of a Prometheus
-	// server in a Thanos environment.
-	//
-	// This section is experimental, it may change significantly without
-	// deprecation notice in any release.
-	//
-	// This is experimental and may change significantly without backward
-	// compatibility in any release.
-	Thanos *ThanosSpec `json:"thanos,omitempty"`
-	// QueryLogFile specifies the file to which PromQL queries are logged.
-	// If the filename has an empty path, e.g. 'query.log', prometheus-operator will mount the file into an
-	// emptyDir volume at `/var/log/prometheus`. If a full path is provided, e.g. /var/log/prometheus/query.log, you must mount a volume
-	// in the specified directory and it must be writable. This is because the prometheus container runs with a read-only root filesystem for security reasons.
-	// Alternatively, the location can be set to a stdout location such as `/dev/stdout` to log
-	// query information to the default Prometheus log stream.
-	// This is only available in versions of Prometheus >= 2.16.0.
-	// For more details, see the Prometheus docs (https://prometheus.io/docs/guides/query-log/)
-	QueryLogFile string `json:"queryLogFile,omitempty"`
-	// AllowOverlappingBlocks enables vertical compaction and vertical query merge in Prometheus.
-	// This is still experimental in Prometheus so it may change in any upcoming release.
-	AllowOverlappingBlocks bool `json:"allowOverlappingBlocks,omitempty"`
-	// Exemplars related settings that are runtime reloadable.
-	// It requires to enable the exemplar storage feature to be effective.
-	Exemplars *Exemplars `json:"exemplars,omitempty"`
-	// Interval between consecutive evaluations. Default: `30s`
-	// +kubebuilder:default:="30s"
-	EvaluationInterval Duration `json:"evaluationInterval,omitempty"`
-	// Enable access to prometheus web admin API. Defaults to the value of `false`.
-	// WARNING: Enabling the admin APIs enables mutating endpoints, to delete data,
-	// shutdown Prometheus, and more. Enabling this should be done with care and the
-	// user is advised to add additional authentication authorization via a proxy to
-	// ensure only clients authorized to perform these actions can do so.
-	// For more information see https://prometheus.io/docs/prometheus/latest/querying/api/#tsdb-admin-apis
-	EnableAdminAPI bool `json:"enableAdminAPI,omitempty"`
-	// Defines the runtime reloadable configuration of the timeseries database
-	// (TSDB).
-	TSDB TSDBSpec `json:"tsdb,omitempty"`
-}
-
-type TSDBSpec struct {
-	// Configures how old an out-of-order/out-of-bounds sample can be w.r.t.
-	// the TSDB max time.
-	// An out-of-order/out-of-bounds sample is ingested into the TSDB as long as
-	// the timestamp of the sample is >= (TSDB.MaxTime - outOfOrderTimeWindow).
-	// Out of order ingestion is an experimental feature and requires
-	// Prometheus >= v2.39.0.
-	OutOfOrderTimeWindow Duration `json:"outOfOrderTimeWindow,omitempty"`
-}
-
-type Exemplars struct {
-	// Maximum number of exemplars stored in memory for all series.
-	// If not set, Prometheus uses its default value.
-	// A value of zero or less than zero disables the storage.
-	MaxSize *int64 `json:"maxSize,omitempty"`
-}
-
-// PrometheusRuleExcludeConfig enables users to configure excluded PrometheusRule names and their namespaces
-// to be ignored while enforcing namespace label for alerts and metrics.
+// PrometheusRuleExcludeConfig enables users to configure excluded
+// PrometheusRule names and their namespaces to be ignored while enforcing
+// namespace label for alerts and metrics.
 type PrometheusRuleExcludeConfig struct {
-	// RuleNamespace - namespace of excluded rule
+	// Namespace of the excluded PrometheusRule object.
 	RuleNamespace string `json:"ruleNamespace"`
-	// RuleNamespace - name of excluded rule
+	// Name of the excluded PrometheusRule object.
 	RuleName string `json:"ruleName"`
+}
+
+type ProxyConfig struct {
+	// `proxyURL` defines the HTTP proxy server to use.
+	//
+	// +kubebuilder:validation:Pattern:="^http(s)?://.+$"
+	// +optional
+	ProxyURL *string `json:"proxyUrl,omitempty"`
+	// `noProxy` is a comma-separated string that can contain IPs, CIDR notation, domain names
+	// that should be excluded from proxying. IP and domain names can
+	// contain port numbers.
+	//
+	// It requires Prometheus >= v2.43.0 or Alertmanager >= 0.25.0.
+	// +optional
+	NoProxy *string `json:"noProxy,omitempty"`
+	// Whether to use the proxy configuration defined by environment variables (HTTP_PROXY, HTTPS_PROXY, and NO_PROXY).
+	//
+	// It requires Prometheus >= v2.43.0 or Alertmanager >= 0.25.0.
+	// +optional
+	ProxyFromEnvironment *bool `json:"proxyFromEnvironment,omitempty"`
+	// ProxyConnectHeader optionally specifies headers to send to
+	// proxies during CONNECT requests.
+	//
+	// It requires Prometheus >= v2.43.0 or Alertmanager >= 0.25.0.
+	// +optional
+	// +mapType:=atomic
+	ProxyConnectHeader map[string][]v1.SecretKeySelector `json:"proxyConnectHeader,omitempty"`
 }
 
 // ObjectReference references a PodMonitor, ServiceMonitor, Probe or PrometheusRule object.
@@ -550,14 +109,14 @@ type ObjectReference struct {
 	Group string `json:"group"`
 	// Resource of the referent.
 	// +kubebuilder:validation:Required
-	// +kubebuilder:validation:Enum=prometheusrules;servicemonitors;podmonitors;probes
+	// +kubebuilder:validation:Enum=prometheusrules;servicemonitors;podmonitors;probes;scrapeconfigs
 	Resource string `json:"resource"`
 	// Namespace of the referent.
 	// More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/namespaces/
 	// +kubebuilder:validation:Required
 	// +kubebuilder:validation:MinLength=1
 	Namespace string `json:"namespace"`
-	// Name of the referent. When not set, all resources are matched.
+	// Name of the referent. When not set, all resources in the namespace are matched.
 	// +optional
 	Name string `json:"name,omitempty"`
 }
@@ -574,45 +133,16 @@ type ArbitraryFSAccessThroughSMsConfig struct {
 	Deny bool `json:"deny,omitempty"`
 }
 
-// PrometheusStatus is the most recent observed status of the Prometheus cluster.
-// More info:
-// https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
-type PrometheusStatus struct {
-	// Represents whether any actions on the underlying managed objects are
-	// being performed. Only delete actions will be performed.
-	Paused bool `json:"paused"`
-	// Total number of non-terminated pods targeted by this Prometheus deployment
-	// (their labels match the selector).
-	Replicas int32 `json:"replicas"`
-	// Total number of non-terminated pods targeted by this Prometheus deployment
-	// that have the desired version spec.
-	UpdatedReplicas int32 `json:"updatedReplicas"`
-	// Total number of available pods (ready for at least minReadySeconds)
-	// targeted by this Prometheus deployment.
-	AvailableReplicas int32 `json:"availableReplicas"`
-	// Total number of unavailable pods targeted by this Prometheus deployment.
-	UnavailableReplicas int32 `json:"unavailableReplicas"`
-	// The current state of the Prometheus deployment.
-	// +listType=map
-	// +listMapKey=type
-	// +optional
-	Conditions []PrometheusCondition `json:"conditions,omitempty"`
-	// The list has one entry per shard. Each entry provides a summary of the shard status.
-	// +listType=map
-	// +listMapKey=shardID
-	// +optional
-	ShardStatuses []ShardStatus `json:"shardStatuses,omitempty"`
-}
-
-// PrometheusCondition represents the state of the resources associated with the Prometheus resource.
+// Condition represents the state of the resources associated with the
+// Prometheus, Alertmanager or ThanosRuler resource.
 // +k8s:deepcopy-gen=true
-type PrometheusCondition struct {
+type Condition struct {
 	// Type of the condition being reported.
 	// +required
-	Type PrometheusConditionType `json:"type"`
-	// status of the condition.
+	Type ConditionType `json:"type"`
+	// Status of the condition.
 	// +required
-	Status PrometheusConditionStatus `json:"status"`
+	Status ConditionStatus `json:"status"`
 	// lastTransitionTime is the time of the last update to the current status property.
 	// +required
 	LastTransitionTime metav1.Time `json:"lastTransitionTime"`
@@ -622,80 +152,44 @@ type PrometheusCondition struct {
 	// Human-readable message indicating details for the condition's last transition.
 	// +optional
 	Message string `json:"message,omitempty"`
-	// ObservedGeneration represents the .metadata.generation that the condition was set based upon.
-	// For instance, if .metadata.generation is currently 12, but the .status.conditions[x].observedGeneration is 9, the condition is out of date
-	// with respect to the current state of the instance.
+	// ObservedGeneration represents the .metadata.generation that the
+	// condition was set based upon. For instance, if `.metadata.generation` is
+	// currently 12, but the `.status.conditions[].observedGeneration` is 9, the
+	// condition is out of date with respect to the current state of the
+	// instance.
 	ObservedGeneration int64 `json:"observedGeneration,omitempty"`
 }
 
-type PrometheusConditionType string
+// +kubebuilder:validation:MinLength=1
+type ConditionType string
 
 const (
-	// Available indicates whether enough Prometheus pods are ready to provide
-	// the service.
+	// Available indicates whether enough pods are ready to provide the
+	// service.
 	// The possible status values for this condition type are:
 	// - True: all pods are running and ready, the service is fully available.
 	// - Degraded: some pods aren't ready, the service is partially available.
 	// - False: no pods are running, the service is totally unavailable.
 	// - Unknown: the operator couldn't determine the condition status.
-	PrometheusAvailable PrometheusConditionType = "Available"
+	Available ConditionType = "Available"
 	// Reconciled indicates whether the operator has reconciled the state of
-	// the underlying resources with the Prometheus object spec.
+	// the underlying resources with the object's spec.
 	// The possible status values for this condition type are:
 	// - True: the reconciliation was successful.
 	// - False: the reconciliation failed.
 	// - Unknown: the operator couldn't determine the condition status.
-	PrometheusReconciled PrometheusConditionType = "Reconciled"
+	Reconciled ConditionType = "Reconciled"
 )
 
-type PrometheusConditionStatus string
+// +kubebuilder:validation:MinLength=1
+type ConditionStatus string
 
 const (
-	PrometheusConditionTrue     PrometheusConditionStatus = "True"
-	PrometheusConditionDegraded PrometheusConditionStatus = "Degraded"
-	PrometheusConditionFalse    PrometheusConditionStatus = "False"
-	PrometheusConditionUnknown  PrometheusConditionStatus = "Unknown"
+	ConditionTrue     ConditionStatus = "True"
+	ConditionDegraded ConditionStatus = "Degraded"
+	ConditionFalse    ConditionStatus = "False"
+	ConditionUnknown  ConditionStatus = "Unknown"
 )
-
-type ShardStatus struct {
-	// Identifier of the shard.
-	// +required
-	ShardID string `json:"shardID"`
-	// Total number of pods targeted by this shard.
-	Replicas int32 `json:"replicas"`
-	// Total number of non-terminated pods targeted by this shard
-	// that have the desired spec.
-	UpdatedReplicas int32 `json:"updatedReplicas"`
-	// Total number of available pods (ready for at least minReadySeconds)
-	// targeted by this shard.
-	AvailableReplicas int32 `json:"availableReplicas"`
-	// Total number of unavailable pods targeted by this shard.
-	UnavailableReplicas int32 `json:"unavailableReplicas"`
-}
-
-// AlertingSpec defines parameters for alerting configuration of Prometheus servers.
-type AlertingSpec struct {
-	// AlertmanagerEndpoints Prometheus should fire alerts against.
-	Alertmanagers []AlertmanagerEndpoints `json:"alertmanagers"`
-}
-
-// StorageSpec defines the configured storage for a group Prometheus servers.
-// If no storage option is specified, then by default an [EmptyDir](https://kubernetes.io/docs/concepts/storage/volumes/#emptydir) will be used.
-// If multiple storage options are specified, priority will be given as follows: EmptyDir, Ephemeral, and lastly VolumeClaimTemplate.
-type StorageSpec struct {
-	// Deprecated: subPath usage will be disabled by default in a future release, this option will become unnecessary.
-	// DisableMountSubPath allows to remove any subPath usage in volume mounts.
-	DisableMountSubPath bool `json:"disableMountSubPath,omitempty"`
-	// EmptyDirVolumeSource to be used by the Prometheus StatefulSets. If specified, used in place of any volumeClaimTemplate. More
-	// info: https://kubernetes.io/docs/concepts/storage/volumes/#emptydir
-	EmptyDir *v1.EmptyDirVolumeSource `json:"emptyDir,omitempty"`
-	// EphemeralVolumeSource to be used by the Prometheus StatefulSets.
-	// This is a beta field in k8s 1.21, for lower versions, starting with k8s 1.19, it requires enabling the GenericEphemeralVolume feature gate.
-	// More info: https://kubernetes.io/docs/concepts/storage/ephemeral-volumes/#generic-ephemeral-volumes
-	Ephemeral *v1.EphemeralVolumeSource `json:"ephemeral,omitempty"`
-	// A PVC spec to be used by the Prometheus StatefulSets.
-	VolumeClaimTemplate EmbeddedPersistentVolumeClaim `json:"volumeClaimTemplate,omitempty"`
-}
 
 // EmbeddedPersistentVolumeClaim is an embedded version of k8s.io/api/core/v1.PersistentVolumeClaim.
 // It contains TypeMeta and a reduced ObjectMeta.
@@ -705,15 +199,13 @@ type EmbeddedPersistentVolumeClaim struct {
 	// EmbeddedMetadata contains metadata relevant to an EmbeddedResource.
 	EmbeddedObjectMetadata `json:"metadata,omitempty" protobuf:"bytes,1,opt,name=metadata"`
 
-	// Spec defines the desired characteristics of a volume requested by a pod author.
+	// Defines the desired characteristics of a volume requested by a pod author.
 	// More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes#persistentvolumeclaims
 	// +optional
 	Spec v1.PersistentVolumeClaimSpec `json:"spec,omitempty" protobuf:"bytes,2,opt,name=spec"`
 
-	// Status represents the current information/status of a persistent volume claim.
-	// Read-only.
-	// More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes#persistentvolumeclaims
 	// +optional
+	// Deprecated: this field is never set.
 	Status v1.PersistentVolumeClaimStatus `json:"status,omitempty" protobuf:"bytes,3,opt,name=status"`
 }
 
@@ -744,30 +236,6 @@ type EmbeddedObjectMetadata struct {
 	Annotations map[string]string `json:"annotations,omitempty" protobuf:"bytes,12,rep,name=annotations"`
 }
 
-// QuerySpec defines the query command line flags when starting Prometheus.
-type QuerySpec struct {
-	// The delta difference allowed for retrieving metrics during expression evaluations.
-	LookbackDelta *string `json:"lookbackDelta,omitempty"`
-	// Number of concurrent queries that can be run at once.
-	MaxConcurrency *int32 `json:"maxConcurrency,omitempty"`
-	// Maximum number of samples a single query can load into memory. Note that queries will fail if they would load more samples than this into memory, so this also limits the number of samples a query can return.
-	MaxSamples *int32 `json:"maxSamples,omitempty"`
-	// Maximum time a query may take before being aborted.
-	Timeout *Duration `json:"timeout,omitempty"`
-}
-
-// PrometheusWebSpec defines the web command line flags when starting Prometheus.
-type PrometheusWebSpec struct {
-	WebConfigFileFields `json:",inline"`
-	// The prometheus web page title
-	PageTitle *string `json:"pageTitle,omitempty"`
-}
-
-// AlertmanagerWebSpec defines the web command line flags when starting Alertmanager.
-type AlertmanagerWebSpec struct {
-	WebConfigFileFields `json:",inline"`
-}
-
 // WebConfigFileFields defines the file content for --web.config.file flag.
 // +k8s:deepcopy-gen=true
 type WebConfigFileFields struct {
@@ -795,12 +263,12 @@ type WebHTTPHeaders struct {
 	// Set the X-Frame-Options header to HTTP responses.
 	// Unset if blank. Accepted values are deny and sameorigin.
 	// https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Frame-Options
-	// +kubebuilder:validation:Enum="";Deny;SameOrigin
+	//+kubebuilder:validation:Enum="";Deny;SameOrigin
 	XFrameOptions string `json:"xFrameOptions,omitempty"`
 	// Set the X-Content-Type-Options header to HTTP responses.
 	// Unset if blank. Accepted value is nosniff.
 	// https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Content-Type-Options
-	// +kubebuilder:validation:Enum="";NoSniff
+	//+kubebuilder:validation:Enum="";NoSniff
 	XContentTypeOptions string `json:"xContentTypeOptions,omitempty"`
 	// Set the X-XSS-Protection header to all responses.
 	// Unset if blank.
@@ -817,673 +285,316 @@ type WebHTTPHeaders struct {
 
 // WebTLSConfig defines the TLS parameters for HTTPS.
 type WebTLSConfig struct {
-	// Secret containing the TLS key for the server.
-	KeySecret v1.SecretKeySelector `json:"keySecret"`
-	// Contains the TLS certificate for the server.
-	Cert SecretOrConfigMap `json:"cert"`
-	// Server policy for client authentication. Maps to ClientAuth Policies.
+	// Secret or ConfigMap containing the TLS certificate for the web server.
+	//
+	// Either `keySecret` or `keyFile` must be defined.
+	//
+	// It is mutually exclusive with `certFile`.
+	//
+	// +optional
+	Cert SecretOrConfigMap `json:"cert,omitempty"`
+	// Path to the TLS certificate file in the container for the web server.
+	//
+	// Either `keySecret` or `keyFile` must be defined.
+	//
+	// It is mutually exclusive with `cert`.
+	//
+	// +optional
+	CertFile *string `json:"certFile,omitempty"`
+
+	// Secret containing the TLS private key for the web server.
+	//
+	// Either `cert` or `certFile` must be defined.
+	//
+	// It is mutually exclusive with `keyFile`.
+	//
+	// +optional
+	KeySecret v1.SecretKeySelector `json:"keySecret,omitempty"`
+	// Path to the TLS private key file in the container for the web server.
+	//
+	// If defined, either `cert` or `certFile` must be defined.
+	//
+	// It is mutually exclusive with `keySecret`.
+	//
+	// +optional
+	KeyFile *string `json:"keyFile,omitempty"`
+
+	// Secret or ConfigMap containing the CA certificate for client certificate
+	// authentication to the server.
+	//
+	// It is mutually exclusive with `clientCAFile`.
+	//
+	// +optional
+	ClientCA SecretOrConfigMap `json:"client_ca,omitempty"`
+	// Path to the CA certificate file for client certificate authentication to
+	// the server.
+	//
+	// It is mutually exclusive with `client_ca`.
+	//
+	// +optional
+	ClientCAFile *string `json:"clientCAFile,omitempty"`
+	// The server policy for client TLS authentication.
+	//
 	// For more detail on clientAuth options:
 	// https://golang.org/pkg/crypto/tls/#ClientAuthType
-	ClientAuthType string `json:"clientAuthType,omitempty"`
-	// Contains the CA certificate for client certificate authentication to the server.
-	ClientCA SecretOrConfigMap `json:"client_ca,omitempty"`
-	// Minimum TLS version that is acceptable. Defaults to TLS12.
-	MinVersion string `json:"minVersion,omitempty"`
-	// Maximum TLS version that is acceptable. Defaults to TLS13.
-	MaxVersion string `json:"maxVersion,omitempty"`
-	// List of supported cipher suites for TLS versions up to TLS 1.2. If empty,
-	// Go default cipher suites are used. Available cipher suites are documented
-	// in the go documentation: https://golang.org/pkg/crypto/tls/#pkg-constants
+	//
+	// +optional
+	ClientAuthType *string `json:"clientAuthType,omitempty"`
+
+	// Minimum TLS version that is acceptable.
+	//
+	// +optional
+	MinVersion *string `json:"minVersion,omitempty"`
+	// Maximum TLS version that is acceptable.
+	//
+	// +optional
+	MaxVersion *string `json:"maxVersion,omitempty"`
+
+	// List of supported cipher suites for TLS versions up to TLS 1.2.
+	//
+	// If not defined, the Go default cipher suites are used.
+	// Available cipher suites are documented in the Go documentation:
+	// https://golang.org/pkg/crypto/tls/#pkg-constants
+	//
+	// +optional
 	CipherSuites []string `json:"cipherSuites,omitempty"`
-	// Controls whether the server selects the
-	// client's most preferred cipher suite, or the server's most preferred
-	// cipher suite. If true then the server's preference, as expressed in
+
+	// Controls whether the server selects the client's most preferred cipher
+	// suite, or the server's most preferred cipher suite.
+	//
+	// If true then the server's preference, as expressed in
 	// the order of elements in cipherSuites, is used.
+	//
+	// +optional
 	PreferServerCipherSuites *bool `json:"preferServerCipherSuites,omitempty"`
+
 	// Elliptic curves that will be used in an ECDHE handshake, in preference
-	// order. Available curves are documented in the go documentation:
+	// order.
+	//
+	// Available curves are documented in the Go documentation:
 	// https://golang.org/pkg/crypto/tls/#CurveID
+	//
+	// +optional
 	CurvePreferences []string `json:"curvePreferences,omitempty"`
 }
 
-// ThanosSpec defines parameters for a Prometheus server within a Thanos deployment.
-type ThanosSpec struct {
-	// Image if specified has precedence over baseImage, tag and sha
-	// combinations. Specifying the version is still necessary to ensure the
-	// Prometheus Operator knows what version of Thanos is being
-	// configured.
-	Image *string `json:"image,omitempty"`
-	// Version describes the version of Thanos to use.
-	Version *string `json:"version,omitempty"`
-	// Tag of Thanos sidecar container image to be deployed. Defaults to the value of `version`.
-	// Version is ignored if Tag is set.
-	// Deprecated: use 'image' instead.  The image tag can be specified
-	// as part of the image URL.
-	Tag *string `json:"tag,omitempty"`
-	// SHA of Thanos container image to be deployed. Defaults to the value of `version`.
-	// Similar to a tag, but the SHA explicitly deploys an immutable container image.
-	// Version and Tag are ignored if SHA is set.
-	// Deprecated: use 'image' instead.  The image digest can be specified
-	// as part of the image URL.
-	SHA *string `json:"sha,omitempty"`
-	// Thanos base image if other than default.
-	// Deprecated: use 'image' instead
-	BaseImage *string `json:"baseImage,omitempty"`
-	// Resources defines the resource requirements for the Thanos sidecar.
-	// If not provided, no requests/limits will be set
-	Resources v1.ResourceRequirements `json:"resources,omitempty"`
-	// ObjectStorageConfig configures object storage in Thanos.
-	// Alternative to ObjectStorageConfigFile, and lower order priority.
-	ObjectStorageConfig *v1.SecretKeySelector `json:"objectStorageConfig,omitempty"`
-	// ObjectStorageConfigFile specifies the path of the object storage configuration file.
-	// When used alongside with ObjectStorageConfig, ObjectStorageConfigFile takes precedence.
-	ObjectStorageConfigFile *string `json:"objectStorageConfigFile,omitempty"`
-	// If true, the Thanos sidecar listens on the loopback interface
-	// for the HTTP and gRPC endpoints.
-	// It takes precedence over `grpcListenLocal` and `httpListenLocal`.
-	// Deprecated: use `grpcListenLocal` and `httpListenLocal` instead.
-	ListenLocal bool `json:"listenLocal,omitempty"`
-	// If true, the Thanos sidecar listens on the loopback interface
-	// for the gRPC endpoints.
-	// It has no effect if `listenLocal` is true.
-	GRPCListenLocal bool `json:"grpcListenLocal,omitempty"`
-	// If true, the Thanos sidecar listens on the loopback interface
-	// for the HTTP endpoints.
-	// It has no effect if `listenLocal` is true.
-	HTTPListenLocal bool `json:"httpListenLocal,omitempty"`
-	// TracingConfig configures tracing in Thanos. This is an experimental feature, it may change in any upcoming release in a breaking way.
-	TracingConfig *v1.SecretKeySelector `json:"tracingConfig,omitempty"`
-	// TracingConfig specifies the path of the tracing configuration file.
-	// When used alongside with TracingConfig, TracingConfigFile takes precedence.
-	TracingConfigFile string `json:"tracingConfigFile,omitempty"`
-	// GRPCServerTLSConfig configures the TLS parameters for the gRPC server
-	// providing the StoreAPI.
-	// Note: Currently only the CAFile, CertFile, and KeyFile fields are supported.
-	// Maps to the '--grpc-server-tls-*' CLI args.
-	GRPCServerTLSConfig *TLSConfig `json:"grpcServerTlsConfig,omitempty"`
-	// LogLevel for Thanos sidecar to be configured with.
-	// +kubebuilder:validation:Enum="";debug;info;warn;error
-	LogLevel string `json:"logLevel,omitempty"`
-	// LogFormat for Thanos sidecar to be configured with.
-	// +kubebuilder:validation:Enum="";logfmt;json
-	LogFormat string `json:"logFormat,omitempty"`
-	// MinTime for Thanos sidecar to be configured with. Option can be a constant time in RFC3339 format or time duration relative to current time, such as -1d or 2h45m. Valid duration units are ms, s, m, h, d, w, y.
-	MinTime string `json:"minTime,omitempty"`
-	// ReadyTimeout is the maximum time Thanos sidecar will wait for Prometheus to start. Eg 10m
-	ReadyTimeout Duration `json:"readyTimeout,omitempty"`
-	// VolumeMounts allows configuration of additional VolumeMounts on the output StatefulSet definition.
-	// VolumeMounts specified will be appended to other VolumeMounts in the thanos-sidecar container.
-	VolumeMounts []v1.VolumeMount `json:"volumeMounts,omitempty"`
-	// AdditionalArgs allows setting additional arguments for the Thanos container.
-	// The arguments are passed as-is to the Thanos container which may cause issues
-	// if they are invalid or not supported the given Thanos version.
-	// In case of an argument conflict (e.g. an argument which is already set by the
-	// operator itself) or when providing an invalid argument the reconciliation will
-	// fail and an error will be logged.
-	AdditionalArgs []Argument `json:"additionalArgs,omitempty"`
-}
-
-// RemoteWriteSpec defines the configuration to write samples from Prometheus
-// to a remote endpoint.
-type RemoteWriteSpec struct {
-	// The URL of the endpoint to send samples to.
-	URL string `json:"url"`
-	// The name of the remote write queue, it must be unique if specified. The
-	// name is used in metrics and logging in order to differentiate queues.
-	// Only valid in Prometheus versions 2.15.0 and newer.
-	Name string `json:"name,omitempty"`
-	// Enables sending of exemplars over remote write. Note that
-	// exemplar-storage itself must be enabled using the enableFeature option
-	// for exemplars to be scraped in the first place.  Only valid in
-	// Prometheus versions 2.27.0 and newer.
-	SendExemplars *bool `json:"sendExemplars,omitempty"`
-	// Timeout for requests to the remote write endpoint.
-	RemoteTimeout Duration `json:"remoteTimeout,omitempty"`
-	// Custom HTTP headers to be sent along with each remote write request.
-	// Be aware that headers that are set by Prometheus itself can't be overwritten.
-	// Only valid in Prometheus versions 2.25.0 and newer.
-	Headers map[string]string `json:"headers,omitempty"`
-	// The list of remote write relabel configurations.
-	WriteRelabelConfigs []RelabelConfig `json:"writeRelabelConfigs,omitempty"`
-	// OAuth2 for the URL. Only valid in Prometheus versions 2.27.0 and newer.
-	OAuth2 *OAuth2 `json:"oauth2,omitempty"`
-	// BasicAuth for the URL.
-	BasicAuth *BasicAuth `json:"basicAuth,omitempty"`
-	// Bearer token for remote write.
-	BearerToken string `json:"bearerToken,omitempty"`
-	// File to read bearer token for remote write.
-	BearerTokenFile string `json:"bearerTokenFile,omitempty"`
-	// Authorization section for remote write
-	Authorization *Authorization `json:"authorization,omitempty"`
-	// Sigv4 allows to configures AWS's Signature Verification 4
-	Sigv4 *Sigv4 `json:"sigv4,omitempty"`
-	// TLS Config to use for remote write.
-	TLSConfig *TLSConfig `json:"tlsConfig,omitempty"`
-	// Optional ProxyURL.
-	ProxyURL string `json:"proxyUrl,omitempty"`
-	// QueueConfig allows tuning of the remote write queue parameters.
-	QueueConfig *QueueConfig `json:"queueConfig,omitempty"`
-	// MetadataConfig configures the sending of series metadata to the remote storage.
-	MetadataConfig *MetadataConfig `json:"metadataConfig,omitempty"`
-}
-
-// QueueConfig allows the tuning of remote write's queue_config parameters.
-// This object is referenced in the RemoteWriteSpec object.
-type QueueConfig struct {
-	// Capacity is the number of samples to buffer per shard before we start dropping them.
-	Capacity int `json:"capacity,omitempty"`
-	// MinShards is the minimum number of shards, i.e. amount of concurrency.
-	MinShards int `json:"minShards,omitempty"`
-	// MaxShards is the maximum number of shards, i.e. amount of concurrency.
-	MaxShards int `json:"maxShards,omitempty"`
-	// MaxSamplesPerSend is the maximum number of samples per send.
-	MaxSamplesPerSend int `json:"maxSamplesPerSend,omitempty"`
-	// BatchSendDeadline is the maximum time a sample will wait in buffer.
-	BatchSendDeadline string `json:"batchSendDeadline,omitempty"`
-	// MaxRetries is the maximum number of times to retry a batch on recoverable errors.
-	MaxRetries int `json:"maxRetries,omitempty"`
-	// MinBackoff is the initial retry delay. Gets doubled for every retry.
-	MinBackoff string `json:"minBackoff,omitempty"`
-	// MaxBackoff is the maximum retry delay.
-	MaxBackoff string `json:"maxBackoff,omitempty"`
-	// Retry upon receiving a 429 status code from the remote-write storage.
-	// This is experimental feature and might change in the future.
-	RetryOnRateLimit bool `json:"retryOnRateLimit,omitempty"`
-}
-
-// Sigv4 optionally configures AWS's Signature Verification 4 signing process to
-// sign requests. Cannot be set at the same time as basic_auth or authorization.
-type Sigv4 struct {
-	// Region is the AWS region. If blank, the region from the default credentials chain used.
-	Region string `json:"region,omitempty"`
-	// AccessKey is the AWS API key. If blank, the environment variable `AWS_ACCESS_KEY_ID` is used.
-	AccessKey *v1.SecretKeySelector `json:"accessKey,omitempty"`
-	// SecretKey is the AWS API secret. If blank, the environment variable `AWS_SECRET_ACCESS_KEY` is used.
-	SecretKey *v1.SecretKeySelector `json:"secretKey,omitempty"`
-	// Profile is the named AWS profile used to authenticate.
-	Profile string `json:"profile,omitempty"`
-	// RoleArn is the named AWS profile used to authenticate.
-	RoleArn string `json:"roleArn,omitempty"`
-}
-
-// RemoteReadSpec defines the configuration for Prometheus to read back samples
-// from a remote endpoint.
-type RemoteReadSpec struct {
-	// The URL of the endpoint to query from.
-	URL string `json:"url"`
-	// The name of the remote read queue, it must be unique if specified. The name
-	// is used in metrics and logging in order to differentiate read
-	// configurations.  Only valid in Prometheus versions 2.15.0 and newer.
-	Name string `json:"name,omitempty"`
-	// An optional list of equality matchers which have to be present
-	// in a selector to query the remote read endpoint.
-	RequiredMatchers map[string]string `json:"requiredMatchers,omitempty"`
-	// Timeout for requests to the remote read endpoint.
-	RemoteTimeout Duration `json:"remoteTimeout,omitempty"`
-	// Custom HTTP headers to be sent along with each remote read request.
-	// Be aware that headers that are set by Prometheus itself can't be overwritten.
-	// Only valid in Prometheus versions 2.26.0 and newer.
-	Headers map[string]string `json:"headers,omitempty"`
-	// Whether reads should be made for queries for time ranges that
-	// the local storage should have complete data for.
-	ReadRecent bool `json:"readRecent,omitempty"`
-	// BasicAuth for the URL.
-	BasicAuth *BasicAuth `json:"basicAuth,omitempty"`
-	// OAuth2 for the URL. Only valid in Prometheus versions 2.27.0 and newer.
-	OAuth2 *OAuth2 `json:"oauth2,omitempty"`
-	// Bearer token for remote read.
-	BearerToken string `json:"bearerToken,omitempty"`
-	// File to read bearer token for remote read.
-	BearerTokenFile string `json:"bearerTokenFile,omitempty"`
-	// Authorization section for remote read
-	Authorization *Authorization `json:"authorization,omitempty"`
-	// TLS Config to use for remote read.
-	TLSConfig *TLSConfig `json:"tlsConfig,omitempty"`
-	// Optional ProxyURL.
-	ProxyURL string `json:"proxyUrl,omitempty"`
-	// Whether to use the external labels as selectors for the remote read endpoint.
-	// Requires Prometheus v2.34.0 and above.
-	FilterExternalLabels *bool `json:"filterExternalLabels,omitempty"`
-}
-
-// LabelName is a valid Prometheus label name which may only contain ASCII letters, numbers, as well as underscores.
+// LabelName is a valid Prometheus label name which may only contain ASCII
+// letters, numbers, as well as underscores.
+//
 // +kubebuilder:validation:Pattern:="^[a-zA-Z_][a-zA-Z0-9_]*$"
 type LabelName string
 
-// RelabelConfig allows dynamic rewriting of the label set, being applied to samples before ingestion.
-// It defines `<metric_relabel_configs>`-section of Prometheus configuration.
-// More info: https://prometheus.io/docs/prometheus/latest/configuration/configuration/#metric_relabel_configs
-type RelabelConfig struct {
-	// The source labels select values from existing labels. Their content is concatenated
-	// using the configured separator and matched against the configured regular expression
-	// for the replace, keep, and drop actions.
-	SourceLabels []LabelName `json:"sourceLabels,omitempty"`
-	// Separator placed between concatenated source label values. default is ';'.
-	Separator string `json:"separator,omitempty"`
-	// Label to which the resulting value is written in a replace action.
-	// It is mandatory for replace actions. Regex capture groups are available.
-	TargetLabel string `json:"targetLabel,omitempty"`
-	// Regular expression against which the extracted value is matched. Default is '(.*)'
-	Regex string `json:"regex,omitempty"`
-	// Modulus to take of the hash of the source label values.
-	Modulus uint64 `json:"modulus,omitempty"`
-	// Replacement value against which a regex replace is performed if the
-	// regular expression matches. Regex capture groups are available. Default is '$1'
-	Replacement string `json:"replacement,omitempty"`
-	// Action to perform based on regex matching. Default is 'replace'.
-	// uppercase and lowercase actions require Prometheus >= 2.36.
-	// +kubebuilder:validation:Enum=replace;Replace;keep;Keep;drop;Drop;hashmod;HashMod;labelmap;LabelMap;labeldrop;LabelDrop;labelkeep;LabelKeep;lowercase;Lowercase;uppercase;Uppercase
-	// +kubebuilder:default=replace
-	Action string `json:"action,omitempty"`
-}
-
-// APIServerConfig defines a host and auth methods to access apiserver.
-// More info: https://prometheus.io/docs/prometheus/latest/configuration/configuration/#kubernetes_sd_config
-type APIServerConfig struct {
-	// Host of apiserver.
-	// A valid string consisting of a hostname or IP followed by an optional port number
-	Host string `json:"host"`
-	// BasicAuth allow an endpoint to authenticate over basic authentication
-	BasicAuth *BasicAuth `json:"basicAuth,omitempty"`
-	// Bearer token for accessing apiserver.
-	BearerToken string `json:"bearerToken,omitempty"`
-	// File to read bearer token for accessing apiserver.
-	BearerTokenFile string `json:"bearerTokenFile,omitempty"`
-	// TLS Config to use for accessing apiserver.
-	TLSConfig *TLSConfig `json:"tlsConfig,omitempty"`
-	// Authorization section for accessing apiserver
-	Authorization *Authorization `json:"authorization,omitempty"`
-}
-
-// AlertmanagerEndpoints defines a selection of a single Endpoints object
-// containing alertmanager IPs to fire alerts against.
-type AlertmanagerEndpoints struct {
-	// Namespace of Endpoints object.
-	Namespace string `json:"namespace"`
-	// Name of Endpoints object in Namespace.
-	Name string `json:"name"`
-	// Port the Alertmanager API is exposed on.
-	Port intstr.IntOrString `json:"port"`
-	// Scheme to use when firing alerts.
-	Scheme string `json:"scheme,omitempty"`
-	// Prefix for the HTTP path alerts are pushed to.
-	PathPrefix string `json:"pathPrefix,omitempty"`
-	// TLS Config to use for alertmanager connection.
-	TLSConfig *TLSConfig `json:"tlsConfig,omitempty"`
-	// BearerTokenFile to read from filesystem to use when authenticating to
-	// Alertmanager.
-	BearerTokenFile string `json:"bearerTokenFile,omitempty"`
-	// Authorization section for this alertmanager endpoint
-	Authorization *SafeAuthorization `json:"authorization,omitempty"`
-	// Version of the Alertmanager API that Prometheus uses to send alerts. It
-	// can be "v1" or "v2".
-	APIVersion string `json:"apiVersion,omitempty"`
-	// Timeout is a per-target Alertmanager timeout when pushing alerts.
-	Timeout *Duration `json:"timeout,omitempty"`
-	// Whether to enable HTTP2.
-	EnableHttp2 *bool `json:"enableHttp2,omitempty"`
-}
-
-// +kubebuilder:object:root=true
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +genclient
-// +kubebuilder:resource:categories="prometheus-operator",shortName="smon"
-
-// ServiceMonitor defines monitoring for a set of services.
-type ServiceMonitor struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of desired Service selection for target discovery by
-	// Prometheus.
-	Spec ServiceMonitorSpec `json:"spec"`
-}
-
-// ServiceMonitorSpec contains specification parameters for a ServiceMonitor.
-type ServiceMonitorSpec struct {
-	// JobLabel selects the label from the associated Kubernetes service which will be used as the `job` label for all metrics.
-	//
-	// For example:
-	// If in `ServiceMonitor.spec.jobLabel: foo` and in `Service.metadata.labels.foo: bar`,
-	// then the `job="bar"` label is added to all metrics.
-	//
-	// If the value of this field is empty or if the label doesn't exist for the given Service, the `job` label of the metrics defaults to the name of the Kubernetes Service.
-	JobLabel string `json:"jobLabel,omitempty"`
-	// TargetLabels transfers labels from the Kubernetes `Service` onto the created metrics.
-	TargetLabels []string `json:"targetLabels,omitempty"`
-	// PodTargetLabels transfers labels on the Kubernetes `Pod` onto the created metrics.
-	PodTargetLabels []string `json:"podTargetLabels,omitempty"`
-	// A list of endpoints allowed as part of this ServiceMonitor.
-	Endpoints []Endpoint `json:"endpoints"`
-	// Selector to select Endpoints objects.
-	Selector metav1.LabelSelector `json:"selector"`
-	// Selector to select which namespaces the Kubernetes Endpoints objects are discovered from.
-	NamespaceSelector NamespaceSelector `json:"namespaceSelector,omitempty"`
-	// SampleLimit defines per-scrape limit on number of scraped samples that will be accepted.
-	SampleLimit uint64 `json:"sampleLimit,omitempty"`
-	// TargetLimit defines a limit on the number of scraped targets that will be accepted.
-	TargetLimit uint64 `json:"targetLimit,omitempty"`
-	// Per-scrape limit on number of labels that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelLimit uint64 `json:"labelLimit,omitempty"`
-	// Per-scrape limit on length of labels name that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelNameLengthLimit uint64 `json:"labelNameLengthLimit,omitempty"`
-	// Per-scrape limit on length of labels value that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelValueLengthLimit uint64 `json:"labelValueLengthLimit,omitempty"`
-	// Attaches node metadata to discovered targets.
-	// Requires Prometheus v2.37.0 and above.
-	AttachMetadata *AttachMetadata `json:"attachMetadata,omitempty"`
-}
-
-// Endpoint defines a scrapeable endpoint serving Prometheus metrics.
+// Endpoint defines an endpoint serving Prometheus metrics to be scraped by
+// Prometheus.
 type Endpoint struct {
-	// Name of the service port this endpoint refers to. Mutually exclusive with targetPort.
+	// Name of the Service port which this endpoint refers to.
+	//
+	// It takes precedence over `targetPort`.
 	Port string `json:"port,omitempty"`
-	// Name or number of the target port of the Pod behind the Service, the port must be specified with container port property. Mutually exclusive with port.
+
+	// Name or number of the target port of the `Pod` object behind the
+	// Service. The port must be specified with the container's port property.
+	//
+	// +optional
 	TargetPort *intstr.IntOrString `json:"targetPort,omitempty"`
-	// HTTP path to scrape for metrics.
+
+	// HTTP path from which to scrape for metrics.
+	//
 	// If empty, Prometheus uses the default value (e.g. `/metrics`).
 	Path string `json:"path,omitempty"`
+
 	// HTTP scheme to use for scraping.
+	//
+	// `http` and `https` are the expected values unless you rewrite the
+	// `__scheme__` label via relabeling.
+	//
+	// If empty, Prometheus uses the default value `http`.
+	//
+	// +kubebuilder:validation:Enum=http;https
 	Scheme string `json:"scheme,omitempty"`
-	// Optional HTTP URL parameters
+
+	// params define optional HTTP URL parameters.
 	Params map[string][]string `json:"params,omitempty"`
-	// Interval at which metrics should be scraped
-	// If not specified Prometheus' global scrape interval is used.
+
+	// Interval at which Prometheus scrapes the metrics from the target.
+	//
+	// If empty, Prometheus uses the global scrape interval.
 	Interval Duration `json:"interval,omitempty"`
-	// Timeout after which the scrape is ended
-	// If not specified, the Prometheus global scrape timeout is used unless it is less than `Interval` in which the latter is used.
+
+	// Timeout after which Prometheus considers the scrape to be failed.
+	//
+	// If empty, Prometheus uses the global scrape timeout unless it is less
+	// than the target's scrape interval value in which the latter is used.
+	// The value cannot be greater than the scrape interval otherwise the operator will reject the resource.
 	ScrapeTimeout Duration `json:"scrapeTimeout,omitempty"`
-	// TLS configuration to use when scraping the endpoint
+
+	// TLS configuration to use when scraping the target.
+	//
+	// +optional
 	TLSConfig *TLSConfig `json:"tlsConfig,omitempty"`
-	// File to read bearer token for scraping targets.
+
+	// File to read bearer token for scraping the target.
+	//
+	// Deprecated: use `authorization` instead.
 	BearerTokenFile string `json:"bearerTokenFile,omitempty"`
-	// Secret to mount to read bearer token for scraping targets. The secret
-	// needs to be in the same namespace as the service monitor and accessible by
-	// the Prometheus Operator.
-	BearerTokenSecret v1.SecretKeySelector `json:"bearerTokenSecret,omitempty"`
-	// Authorization section for this endpoint
+
+	// `bearerTokenSecret` specifies a key of a Secret containing the bearer
+	// token for scraping targets. The secret needs to be in the same namespace
+	// as the ServiceMonitor object and readable by the Prometheus Operator.
+	//
+	// +optional
+	//
+	// Deprecated: use `authorization` instead.
+	BearerTokenSecret *v1.SecretKeySelector `json:"bearerTokenSecret,omitempty"`
+
+	// `authorization` configures the Authorization header credentials to use when
+	// scraping the target.
+	//
+	// Cannot be set at the same time as `basicAuth`, or `oauth2`.
+	//
+	// +optional
 	Authorization *SafeAuthorization `json:"authorization,omitempty"`
-	// HonorLabels chooses the metric's labels on collisions with target labels.
+
+	// When true, `honorLabels` preserves the metric's labels when they collide
+	// with the target's labels.
 	HonorLabels bool `json:"honorLabels,omitempty"`
-	// HonorTimestamps controls whether Prometheus respects the timestamps present in scraped data.
+
+	// `honorTimestamps` controls whether Prometheus preserves the timestamps
+	// when exposed by the target.
+	//
+	// +optional
 	HonorTimestamps *bool `json:"honorTimestamps,omitempty"`
-	// BasicAuth allow an endpoint to authenticate over basic authentication
-	// More info: https://prometheus.io/docs/operating/configuration/#endpoints
+
+	// `trackTimestampsStaleness` defines whether Prometheus tracks staleness of
+	// the metrics that have an explicit timestamp present in scraped data.
+	// Has no effect if `honorTimestamps` is false.
+	//
+	// It requires Prometheus >= v2.48.0.
+	//
+	// +optional
+	TrackTimestampsStaleness *bool `json:"trackTimestampsStaleness,omitempty"`
+
+	// `basicAuth` configures the Basic Authentication credentials to use when
+	// scraping the target.
+	//
+	// Cannot be set at the same time as `authorization`, or `oauth2`.
+	//
+	// +optional
 	BasicAuth *BasicAuth `json:"basicAuth,omitempty"`
-	// OAuth2 for the URL. Only valid in Prometheus versions 2.27.0 and newer.
+
+	// `oauth2` configures the OAuth2 settings to use when scraping the target.
+	//
+	// It requires Prometheus >= 2.27.0.
+	//
+	// Cannot be set at the same time as `authorization`, or `basicAuth`.
+	//
+	// +optional
 	OAuth2 *OAuth2 `json:"oauth2,omitempty"`
-	// MetricRelabelConfigs to apply to samples before ingestion.
-	MetricRelabelConfigs []*RelabelConfig `json:"metricRelabelings,omitempty"`
-	// RelabelConfigs to apply to samples before scraping.
-	// Prometheus Operator automatically adds relabelings for a few standard Kubernetes fields.
+
+	// `metricRelabelings` configures the relabeling rules to apply to the
+	// samples before ingestion.
+	//
+	// +optional
+	MetricRelabelConfigs []RelabelConfig `json:"metricRelabelings,omitempty"`
+
+	// `relabelings` configures the relabeling rules to apply the target's
+	// metadata labels.
+	//
+	// The Operator automatically adds relabelings for a few standard Kubernetes fields.
+	//
 	// The original scrape job's name is available via the `__tmp_prometheus_job_name` label.
+	//
 	// More info: https://prometheus.io/docs/prometheus/latest/configuration/configuration/#relabel_config
-	RelabelConfigs []*RelabelConfig `json:"relabelings,omitempty"`
-	// ProxyURL eg http://proxyserver:2195 Directs scrapes to proxy through this endpoint.
+	//
+	// +optional
+	RelabelConfigs []RelabelConfig `json:"relabelings,omitempty"`
+
+	// `proxyURL` configures the HTTP Proxy URL (e.g.
+	// "http://proxyserver:2195") to go through when scraping the target.
+	//
+	// +optional
 	ProxyURL *string `json:"proxyUrl,omitempty"`
-	// FollowRedirects configures whether scrape requests follow HTTP 3xx redirects.
+
+	// `followRedirects` defines whether the scrape requests should follow HTTP
+	// 3xx redirects.
+	//
+	// +optional
 	FollowRedirects *bool `json:"followRedirects,omitempty"`
-	// Whether to enable HTTP2.
+
+	// `enableHttp2` can be used to disable HTTP2 when scraping the target.
+	//
+	// +optional
 	EnableHttp2 *bool `json:"enableHttp2,omitempty"`
-	// Drop pods that are not running. (Failed, Succeeded). Enabled by default.
+
+	// When true, the pods which are not running (e.g. either in Failed or
+	// Succeeded state) are dropped during the target discovery.
+	//
+	// If unset, the filtering is enabled.
+	//
 	// More info: https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/#pod-phase
+	//
+	// +optional
 	FilterRunning *bool `json:"filterRunning,omitempty"`
-}
-
-// +kubebuilder:object:root=true
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +genclient
-// +kubebuilder:resource:categories="prometheus-operator",shortName="pmon"
-
-// PodMonitor defines monitoring for a set of pods.
-type PodMonitor struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of desired Pod selection for target discovery by Prometheus.
-	Spec PodMonitorSpec `json:"spec"`
-}
-
-// PodMonitorSpec contains specification parameters for a PodMonitor.
-type PodMonitorSpec struct {
-	// The label to use to retrieve the job name from.
-	JobLabel string `json:"jobLabel,omitempty"`
-	// PodTargetLabels transfers labels on the Kubernetes Pod onto the target.
-	PodTargetLabels []string `json:"podTargetLabels,omitempty"`
-	// A list of endpoints allowed as part of this PodMonitor.
-	PodMetricsEndpoints []PodMetricsEndpoint `json:"podMetricsEndpoints"`
-	// Selector to select Pod objects.
-	Selector metav1.LabelSelector `json:"selector"`
-	// Selector to select which namespaces the Endpoints objects are discovered from.
-	NamespaceSelector NamespaceSelector `json:"namespaceSelector,omitempty"`
-	// SampleLimit defines per-scrape limit on number of scraped samples that will be accepted.
-	SampleLimit uint64 `json:"sampleLimit,omitempty"`
-	// TargetLimit defines a limit on the number of scraped targets that will be accepted.
-	TargetLimit uint64 `json:"targetLimit,omitempty"`
-	// Per-scrape limit on number of labels that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelLimit uint64 `json:"labelLimit,omitempty"`
-	// Per-scrape limit on length of labels name that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelNameLengthLimit uint64 `json:"labelNameLengthLimit,omitempty"`
-	// Per-scrape limit on length of labels value that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelValueLengthLimit uint64 `json:"labelValueLengthLimit,omitempty"`
-	// Attaches node metadata to discovered targets.
-	// Requires Prometheus v2.35.0 and above.
-	AttachMetadata *AttachMetadata `json:"attachMetadata,omitempty"`
 }
 
 type AttachMetadata struct {
-	// When set to true, Prometheus must have permissions to get Nodes.
-	Node bool `json:"node,omitempty"`
+	// When set to true, Prometheus attaches node metadata to the discovered
+	// targets.
+	//
+	// The Prometheus service account must have the `list` and `watch`
+	// permissions on the `Nodes` objects.
+	//
+	// +optional
+	Node *bool `json:"node,omitempty"`
 }
 
-// PodMetricsEndpoint defines a scrapeable endpoint of a Kubernetes Pod serving Prometheus metrics.
-type PodMetricsEndpoint struct {
-	// Name of the pod port this endpoint refers to. Mutually exclusive with targetPort.
-	Port string `json:"port,omitempty"`
-	// Deprecated: Use 'port' instead.
-	TargetPort *intstr.IntOrString `json:"targetPort,omitempty"`
-	// HTTP path to scrape for metrics.
-	// If empty, Prometheus uses the default value (e.g. `/metrics`).
-	Path string `json:"path,omitempty"`
-	// HTTP scheme to use for scraping.
-	Scheme string `json:"scheme,omitempty"`
-	// Optional HTTP URL parameters
-	Params map[string][]string `json:"params,omitempty"`
-	// Interval at which metrics should be scraped
-	// If not specified Prometheus' global scrape interval is used.
-	Interval Duration `json:"interval,omitempty"`
-	// Timeout after which the scrape is ended
-	// If not specified, the Prometheus global scrape interval is used.
-	ScrapeTimeout Duration `json:"scrapeTimeout,omitempty"`
-	// TLS configuration to use when scraping the endpoint.
-	TLSConfig *PodMetricsEndpointTLSConfig `json:"tlsConfig,omitempty"`
-	// Secret to mount to read bearer token for scraping targets. The secret
-	// needs to be in the same namespace as the pod monitor and accessible by
-	// the Prometheus Operator.
-	BearerTokenSecret v1.SecretKeySelector `json:"bearerTokenSecret,omitempty"`
-	// HonorLabels chooses the metric's labels on collisions with target labels.
-	HonorLabels bool `json:"honorLabels,omitempty"`
-	// HonorTimestamps controls whether Prometheus respects the timestamps present in scraped data.
-	HonorTimestamps *bool `json:"honorTimestamps,omitempty"`
-	// BasicAuth allow an endpoint to authenticate over basic authentication.
-	// More info: https://prometheus.io/docs/operating/configuration/#endpoint
-	BasicAuth *BasicAuth `json:"basicAuth,omitempty"`
-	// OAuth2 for the URL. Only valid in Prometheus versions 2.27.0 and newer.
-	OAuth2 *OAuth2 `json:"oauth2,omitempty"`
-	// Authorization section for this endpoint
-	Authorization *SafeAuthorization `json:"authorization,omitempty"`
-	// MetricRelabelConfigs to apply to samples before ingestion.
-	MetricRelabelConfigs []*RelabelConfig `json:"metricRelabelings,omitempty"`
-	// RelabelConfigs to apply to samples before scraping.
-	// Prometheus Operator automatically adds relabelings for a few standard Kubernetes fields.
-	// The original scrape job's name is available via the `__tmp_prometheus_job_name` label.
-	// More info: https://prometheus.io/docs/prometheus/latest/configuration/configuration/#relabel_config
-	RelabelConfigs []*RelabelConfig `json:"relabelings,omitempty"`
-	// ProxyURL eg http://proxyserver:2195 Directs scrapes to proxy through this endpoint.
-	ProxyURL *string `json:"proxyUrl,omitempty"`
-	// FollowRedirects configures whether scrape requests follow HTTP 3xx redirects.
-	FollowRedirects *bool `json:"followRedirects,omitempty"`
-	// Whether to enable HTTP2.
-	EnableHttp2 *bool `json:"enableHttp2,omitempty"`
-	// Drop pods that are not running. (Failed, Succeeded). Enabled by default.
-	// More info: https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/#pod-phase
-	FilterRunning *bool `json:"filterRunning,omitempty"`
-}
-
-// PodMetricsEndpointTLSConfig specifies TLS configuration parameters.
-type PodMetricsEndpointTLSConfig struct {
-	SafeTLSConfig `json:",inline"`
-}
-
-// +kubebuilder:object:root=true
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +genclient
-// +kubebuilder:resource:categories="prometheus-operator",shortName="prb"
-
-// Probe defines monitoring for a set of static targets or ingresses.
-type Probe struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of desired Ingress selection for target discovery by Prometheus.
-	Spec ProbeSpec `json:"spec"`
-}
-
-// ProbeSpec contains specification parameters for a Probe.
-type ProbeSpec struct {
-	// The job name assigned to scraped metrics by default.
-	JobName string `json:"jobName,omitempty"`
-	// Specification for the prober to use for probing targets.
-	// The prober.URL parameter is required. Targets cannot be probed if left empty.
-	ProberSpec ProberSpec `json:"prober,omitempty"`
-	// The module to use for probing specifying how to probe the target.
-	// Example module configuring in the blackbox exporter:
-	// https://github.com/prometheus/blackbox_exporter/blob/master/example.yml
-	Module string `json:"module,omitempty"`
-	// Targets defines a set of static or dynamically discovered targets to probe.
-	Targets ProbeTargets `json:"targets,omitempty"`
-	// Interval at which targets are probed using the configured prober.
-	// If not specified Prometheus' global scrape interval is used.
-	Interval Duration `json:"interval,omitempty"`
-	// Timeout for scraping metrics from the Prometheus exporter.
-	// If not specified, the Prometheus global scrape interval is used.
-	ScrapeTimeout Duration `json:"scrapeTimeout,omitempty"`
-	// TLS configuration to use when scraping the endpoint.
-	TLSConfig *ProbeTLSConfig `json:"tlsConfig,omitempty"`
-	// Secret to mount to read bearer token for scraping targets. The secret
-	// needs to be in the same namespace as the probe and accessible by
-	// the Prometheus Operator.
-	BearerTokenSecret v1.SecretKeySelector `json:"bearerTokenSecret,omitempty"`
-	// BasicAuth allow an endpoint to authenticate over basic authentication.
-	// More info: https://prometheus.io/docs/operating/configuration/#endpoint
-	BasicAuth *BasicAuth `json:"basicAuth,omitempty"`
-	// OAuth2 for the URL. Only valid in Prometheus versions 2.27.0 and newer.
-	OAuth2 *OAuth2 `json:"oauth2,omitempty"`
-	// MetricRelabelConfigs to apply to samples before ingestion.
-	MetricRelabelConfigs []*RelabelConfig `json:"metricRelabelings,omitempty"`
-	// Authorization section for this endpoint
-	Authorization *SafeAuthorization `json:"authorization,omitempty"`
-	// SampleLimit defines per-scrape limit on number of scraped samples that will be accepted.
-	SampleLimit uint64 `json:"sampleLimit,omitempty"`
-	// TargetLimit defines a limit on the number of scraped targets that will be accepted.
-	TargetLimit uint64 `json:"targetLimit,omitempty"`
-	// Per-scrape limit on number of labels that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelLimit uint64 `json:"labelLimit,omitempty"`
-	// Per-scrape limit on length of labels name that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelNameLengthLimit uint64 `json:"labelNameLengthLimit,omitempty"`
-	// Per-scrape limit on length of labels value that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelValueLengthLimit uint64 `json:"labelValueLengthLimit,omitempty"`
-}
-
-// ProbeTargets defines how to discover the probed targets.
-// One of the `staticConfig` or `ingress` must be defined.
-// If both are defined, `staticConfig` takes precedence.
-type ProbeTargets struct {
-	// staticConfig defines the static list of targets to probe and the
-	// relabeling configuration.
-	// If `ingress` is also defined, `staticConfig` takes precedence.
-	// More info: https://prometheus.io/docs/prometheus/latest/configuration/configuration/#static_config.
-	StaticConfig *ProbeTargetStaticConfig `json:"staticConfig,omitempty"`
-	// ingress defines the Ingress objects to probe and the relabeling
-	// configuration.
-	// If `staticConfig` is also defined, `staticConfig` takes precedence.
-	Ingress *ProbeTargetIngress `json:"ingress,omitempty"`
-}
-
-// ProbeTargetStaticConfig defines the set of static targets considered for probing.
-type ProbeTargetStaticConfig struct {
-	// The list of hosts to probe.
-	Targets []string `json:"static,omitempty"`
-	// Labels assigned to all metrics scraped from the targets.
-	Labels map[string]string `json:"labels,omitempty"`
-	// RelabelConfigs to apply to the label set of the targets before it gets
-	// scraped.
-	// More info: https://prometheus.io/docs/prometheus/latest/configuration/configuration/#relabel_config
-	RelabelConfigs []*RelabelConfig `json:"relabelingConfigs,omitempty"`
-}
-
-// ProbeTargetIngress defines the set of Ingress objects considered for probing.
-// The operator configures a target for each host/path combination of each ingress object.
-type ProbeTargetIngress struct {
-	// Selector to select the Ingress objects.
-	Selector metav1.LabelSelector `json:"selector,omitempty"`
-	// From which namespaces to select Ingress objects.
-	NamespaceSelector NamespaceSelector `json:"namespaceSelector,omitempty"`
-	// RelabelConfigs to apply to the label set of the target before it gets
-	// scraped.
-	// The original ingress address is available via the
-	// `__tmp_prometheus_ingress_address` label. It can be used to customize the
-	// probed URL.
-	// The original scrape job's name is available via the `__tmp_prometheus_job_name` label.
-	// More info: https://prometheus.io/docs/prometheus/latest/configuration/configuration/#relabel_config
-	RelabelConfigs []*RelabelConfig `json:"relabelingConfigs,omitempty"`
-}
-
-// ProberSpec contains specification parameters for the Prober used for probing.
-type ProberSpec struct {
-	// Mandatory URL of the prober.
-	URL string `json:"url"`
-	// HTTP scheme to use for scraping.
-	// Defaults to `http`.
-	Scheme string `json:"scheme,omitempty"`
-	// Path to collect metrics from.
-	// Defaults to `/probe`.
-	// +kubebuilder:default:="/probe"
-	Path string `json:"path,omitempty"`
-	// Optional ProxyURL.
-	ProxyURL string `json:"proxyUrl,omitempty"`
-}
-
-// OAuth2 allows an endpoint to authenticate with OAuth2.
-// More info: https://prometheus.io/docs/prometheus/latest/configuration/configuration/#oauth2
+// OAuth2 configures OAuth2 settings.
 type OAuth2 struct {
-	// The secret or configmap containing the OAuth2 client id
+	// `clientId` specifies a key of a Secret or ConfigMap containing the
+	// OAuth2 client's ID.
 	ClientID SecretOrConfigMap `json:"clientId"`
-	// The secret containing the OAuth2 client secret
+
+	// `clientSecret` specifies a key of a Secret containing the OAuth2
+	// client's secret.
 	ClientSecret v1.SecretKeySelector `json:"clientSecret"`
-	// The URL to fetch the token from
+
+	// `tokenURL` configures the URL to fetch the token from.
+	//
 	// +kubebuilder:validation:MinLength=1
 	TokenURL string `json:"tokenUrl"`
-	// OAuth2 scopes used for the token request
+
+	// `scopes` defines the OAuth2 scopes used for the token request.
+	//
+	// +optional.
 	Scopes []string `json:"scopes,omitempty"`
-	// Parameters to append to the token URL
+
+	// `endpointParams` configures the HTTP parameters to append to the token
+	// URL.
+	//
+	// +optional
 	EndpointParams map[string]string `json:"endpointParams,omitempty"`
+
+	// TLS configuration to use when connecting to the OAuth2 server.
+	// It requires Prometheus >= v2.43.0.
+	//
+	// +optional
+	TLSConfig *SafeTLSConfig `json:"tlsConfig,omitempty"`
+
+	// Proxy configuration to use when connecting to the OAuth2 server.
+	// It requires Prometheus >= v2.43.0.
+	//
+	// +optional
+	ProxyConfig `json:",inline"`
 }
 
-// BasicAuth allow an endpoint to authenticate over basic authentication
-// More info: https://prometheus.io/docs/operating/configuration/#endpoints
+// BasicAuth configures HTTP Basic Authentication settings.
 type BasicAuth struct {
-	// The secret in the service monitor namespace that contains the username
-	// for authentication.
+	// `username` specifies a key of a Secret containing the username for
+	// authentication.
 	Username v1.SecretKeySelector `json:"username,omitempty"`
-	// The secret in the service monitor namespace that contains the password
-	// for authentication.
+
+	// `password` specifies a key of a Secret containing the password for
+	// authentication.
 	Password v1.SecretKeySelector `json:"password,omitempty"`
 }
 
@@ -1495,18 +606,46 @@ type SecretOrConfigMap struct {
 	ConfigMap *v1.ConfigMapKeySelector `json:"configMap,omitempty"`
 }
 
+// +kubebuilder:validation:Enum=TLS10;TLS11;TLS12;TLS13
+type TLSVersion string
+
+const (
+	TLSVersion10 TLSVersion = "TLS10"
+	TLSVersion11 TLSVersion = "TLS11"
+	TLSVersion12 TLSVersion = "TLS12"
+	TLSVersion13 TLSVersion = "TLS13"
+)
+
 // SafeTLSConfig specifies safe TLS configuration parameters.
 type SafeTLSConfig struct {
 	// Certificate authority used when verifying server certificates.
 	CA SecretOrConfigMap `json:"ca,omitempty"`
+
 	// Client certificate to present when doing client-authentication.
 	Cert SecretOrConfigMap `json:"cert,omitempty"`
+
 	// Secret containing the client key file for the targets.
 	KeySecret *v1.SecretKeySelector `json:"keySecret,omitempty"`
+
 	// Used to verify the hostname for the targets.
-	ServerName string `json:"serverName,omitempty"`
+	// +optional
+	ServerName *string `json:"serverName,omitempty"`
+
 	// Disable target certificate validation.
-	InsecureSkipVerify bool `json:"insecureSkipVerify,omitempty"`
+	// +optional
+	InsecureSkipVerify *bool `json:"insecureSkipVerify,omitempty"`
+
+	// Minimum acceptable TLS version.
+	//
+	// It requires Prometheus >= v2.35.0.
+	// +optional
+	MinVersion *TLSVersion `json:"minVersion,omitempty"`
+
+	// Maximum acceptable TLS version.
+	//
+	// It requires Prometheus >= v2.41.0.
+	// +optional
+	MaxVersion *TLSVersion `json:"maxVersion,omitempty"`
 }
 
 // TLSConfig extends the safe TLS configuration with file parameters.
@@ -1518,424 +657,6 @@ type TLSConfig struct {
 	CertFile string `json:"certFile,omitempty"`
 	// Path to the client key file in the Prometheus container for the targets.
 	KeyFile string `json:"keyFile,omitempty"`
-}
-
-// +kubebuilder:object:root=true
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
-// ServiceMonitorList is a list of ServiceMonitors.
-type ServiceMonitorList struct {
-	metav1.TypeMeta `json:",inline"`
-	// Standard list metadata
-	// More info: https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#metadata
-	metav1.ListMeta `json:"metadata,omitempty"`
-	// List of ServiceMonitors
-	Items []*ServiceMonitor `json:"items"`
-}
-
-// +kubebuilder:object:root=true
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
-// PodMonitorList is a list of PodMonitors.
-type PodMonitorList struct {
-	metav1.TypeMeta `json:",inline"`
-	// Standard list metadata
-	// More info: https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#metadata
-	metav1.ListMeta `json:"metadata,omitempty"`
-	// List of PodMonitors
-	Items []*PodMonitor `json:"items"`
-}
-
-// +kubebuilder:object:root=true
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
-// ProbeList is a list of Probes.
-type ProbeList struct {
-	metav1.TypeMeta `json:",inline"`
-	// Standard list metadata
-	// More info: https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#metadata
-	metav1.ListMeta `json:"metadata,omitempty"`
-	// List of Probes
-	Items []*Probe `json:"items"`
-}
-
-// +kubebuilder:object:root=true
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
-// PrometheusRuleList is a list of PrometheusRules.
-type PrometheusRuleList struct {
-	metav1.TypeMeta `json:",inline"`
-	// Standard list metadata
-	// More info: https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#metadata
-	metav1.ListMeta `json:"metadata,omitempty"`
-	// List of Rules
-	Items []*PrometheusRule `json:"items"`
-}
-
-// +kubebuilder:object:root=true
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +genclient
-// +kubebuilder:resource:categories="prometheus-operator",shortName="promrule"
-
-// PrometheusRule defines recording and alerting rules for a Prometheus instance
-type PrometheusRule struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of desired alerting rule definitions for Prometheus.
-	Spec PrometheusRuleSpec `json:"spec"`
-}
-
-// PrometheusRuleSpec contains specification parameters for a Rule.
-type PrometheusRuleSpec struct {
-	// Content of Prometheus rule file
-	// +listType=map
-	// +listMapKey=name
-	Groups []RuleGroup `json:"groups,omitempty"`
-}
-
-// RuleGroup and Rule are copied instead of vendored because the
-// upstream Prometheus struct definitions don't have json struct tags.
-
-// RuleGroup is a list of sequentially evaluated recording and alerting rules.
-type RuleGroup struct {
-	// Name of the rule group.
-	// +kubebuilder:validation:MinLength=1
-	Name string `json:"name"`
-	// Interval determines how often rules in the group are evaluated.
-	Interval Duration `json:"interval,omitempty"`
-	// List of alerting and recording rules.
-	Rules []Rule `json:"rules"`
-	// PartialResponseStrategy is only used by ThanosRuler and will
-	// be ignored by Prometheus instances.
-	// More info: https://github.com/thanos-io/thanos/blob/main/docs/components/rule.md#partial-response
-	// +kubebuilder:validation:Pattern="^(?i)(abort|warn)?$"
-	// +kubebuilder:default:=""
-	PartialResponseStrategy string `json:"partial_response_strategy,omitempty"`
-}
-
-// Rule describes an alerting or recording rule
-// See Prometheus documentation: [alerting](https://www.prometheus.io/docs/prometheus/latest/configuration/alerting_rules/) or [recording](https://www.prometheus.io/docs/prometheus/latest/configuration/recording_rules/#recording-rules) rule
-type Rule struct {
-	// Name of the time series to output to. Must be a valid metric name.
-	// Only one of `record` and `alert` must be set.
-	Record string `json:"record,omitempty"`
-	// Name of the alert. Must be a valid label value.
-	// Only one of `record` and `alert` must be set.
-	Alert string `json:"alert,omitempty"`
-	// PromQL expression to evaluate.
-	Expr intstr.IntOrString `json:"expr"`
-	// Alerts are considered firing once they have been returned for this long.
-	For Duration `json:"for,omitempty"`
-	// Labels to add or overwrite.
-	Labels map[string]string `json:"labels,omitempty"`
-	// Annotations to add to each alert.
-	// Only valid for alerting rules.
-	Annotations map[string]string `json:"annotations,omitempty"`
-}
-
-// +kubebuilder:object:root=true
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +genclient
-// +kubebuilder:resource:categories="prometheus-operator",shortName="am"
-// +kubebuilder:printcolumn:name="Version",type="string",JSONPath=".spec.version",description="The version of Alertmanager"
-// +kubebuilder:printcolumn:name="Replicas",type="integer",JSONPath=".spec.replicas",description="The number of desired replicas"
-// +kubebuilder:printcolumn:name="Age",type="date",JSONPath=".metadata.creationTimestamp"
-// +kubebuilder:printcolumn:name="Paused",type="boolean",JSONPath=".status.paused",description="Whether the resource reconciliation is paused or not",priority=1
-
-// Alertmanager describes an Alertmanager cluster.
-type Alertmanager struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of the desired behavior of the Alertmanager cluster. More info:
-	// https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
-	Spec AlertmanagerSpec `json:"spec"`
-	// Most recent observed status of the Alertmanager cluster. Read-only. Not
-	// included when requesting from the apiserver, only from the Prometheus
-	// Operator API itself. More info:
-	// https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
-	Status *AlertmanagerStatus `json:"status,omitempty"`
-}
-
-// AlertmanagerSpec is a specification of the desired behavior of the Alertmanager cluster. More info:
-// https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
-type AlertmanagerSpec struct {
-	// PodMetadata configures Labels and Annotations which are propagated to the alertmanager pods.
-	PodMetadata *EmbeddedObjectMetadata `json:"podMetadata,omitempty"`
-	// Image if specified has precedence over baseImage, tag and sha
-	// combinations. Specifying the version is still necessary to ensure the
-	// Prometheus Operator knows what version of Alertmanager is being
-	// configured.
-	Image *string `json:"image,omitempty"`
-	// Version the cluster should be on.
-	Version string `json:"version,omitempty"`
-	// Tag of Alertmanager container image to be deployed. Defaults to the value of `version`.
-	// Version is ignored if Tag is set.
-	// Deprecated: use 'image' instead.  The image tag can be specified
-	// as part of the image URL.
-	Tag string `json:"tag,omitempty"`
-	// SHA of Alertmanager container image to be deployed. Defaults to the value of `version`.
-	// Similar to a tag, but the SHA explicitly deploys an immutable container image.
-	// Version and Tag are ignored if SHA is set.
-	// Deprecated: use 'image' instead.  The image digest can be specified
-	// as part of the image URL.
-	SHA string `json:"sha,omitempty"`
-	// Base image that is used to deploy pods, without tag.
-	// Deprecated: use 'image' instead
-	BaseImage string `json:"baseImage,omitempty"`
-	// An optional list of references to secrets in the same namespace
-	// to use for pulling prometheus and alertmanager images from registries
-	// see http://kubernetes.io/docs/user-guide/images#specifying-imagepullsecrets-on-a-pod
-	ImagePullSecrets []v1.LocalObjectReference `json:"imagePullSecrets,omitempty"`
-	// Secrets is a list of Secrets in the same namespace as the Alertmanager
-	// object, which shall be mounted into the Alertmanager Pods.
-	// Each Secret is added to the StatefulSet definition as a volume named `secret-<secret-name>`.
-	// The Secrets are mounted into `/etc/alertmanager/secrets/<secret-name>` in the 'alertmanager' container.
-	Secrets []string `json:"secrets,omitempty"`
-	// ConfigMaps is a list of ConfigMaps in the same namespace as the Alertmanager
-	// object, which shall be mounted into the Alertmanager Pods.
-	// Each ConfigMap is added to the StatefulSet definition as a volume named `configmap-<configmap-name>`.
-	// The ConfigMaps are mounted into `/etc/alertmanager/configmaps/<configmap-name>` in the 'alertmanager' container.
-	ConfigMaps []string `json:"configMaps,omitempty"`
-	// ConfigSecret is the name of a Kubernetes Secret in the same namespace as the
-	// Alertmanager object, which contains the configuration for this Alertmanager
-	// instance. If empty, it defaults to `alertmanager-<alertmanager-name>`.
-	//
-	// The Alertmanager configuration should be available under the
-	// `alertmanager.yaml` key. Additional keys from the original secret are
-	// copied to the generated secret and mounted into the
-	// `/etc/alertmanager/config` directory in the `alertmanager` container.
-	//
-	// If either the secret or the `alertmanager.yaml` key is missing, the
-	// operator provisions a minimal Alertmanager configuration with one empty
-	// receiver (effectively dropping alert notifications).
-	ConfigSecret string `json:"configSecret,omitempty"`
-	// Log level for Alertmanager to be configured with.
-	// +kubebuilder:validation:Enum="";debug;info;warn;error
-	LogLevel string `json:"logLevel,omitempty"`
-	// Log format for Alertmanager to be configured with.
-	// +kubebuilder:validation:Enum="";logfmt;json
-	LogFormat string `json:"logFormat,omitempty"`
-	// Size is the expected size of the alertmanager cluster. The controller will
-	// eventually make the size of the running cluster equal to the expected
-	// size.
-	Replicas *int32 `json:"replicas,omitempty"`
-	// Time duration Alertmanager shall retain data for. Default is '120h',
-	// and must match the regular expression `[0-9]+(ms|s|m|h)` (milliseconds seconds minutes hours).
-	// +kubebuilder:default:="120h"
-	Retention GoDuration `json:"retention,omitempty"`
-	// Storage is the definition of how storage will be used by the Alertmanager
-	// instances.
-	Storage *StorageSpec `json:"storage,omitempty"`
-	// Volumes allows configuration of additional volumes on the output StatefulSet definition.
-	// Volumes specified will be appended to other volumes that are generated as a result of
-	// StorageSpec objects.
-	Volumes []v1.Volume `json:"volumes,omitempty"`
-	// VolumeMounts allows configuration of additional VolumeMounts on the output StatefulSet definition.
-	// VolumeMounts specified will be appended to other VolumeMounts in the alertmanager container,
-	// that are generated as a result of StorageSpec objects.
-	VolumeMounts []v1.VolumeMount `json:"volumeMounts,omitempty"`
-	// The external URL the Alertmanager instances will be available under. This is
-	// necessary to generate correct URLs. This is necessary if Alertmanager is not
-	// served from root of a DNS name.
-	ExternalURL string `json:"externalUrl,omitempty"`
-	// The route prefix Alertmanager registers HTTP handlers for. This is useful,
-	// if using ExternalURL and a proxy is rewriting HTTP routes of a request,
-	// and the actual ExternalURL is still true, but the server serves requests
-	// under a different route prefix. For example for use with `kubectl proxy`.
-	RoutePrefix string `json:"routePrefix,omitempty"`
-	// If set to true all actions on the underlying managed objects are not
-	// goint to be performed, except for delete actions.
-	Paused bool `json:"paused,omitempty"`
-	// Define which Nodes the Pods are scheduled on.
-	NodeSelector map[string]string `json:"nodeSelector,omitempty"`
-	// Define resources requests and limits for single Pods.
-	Resources v1.ResourceRequirements `json:"resources,omitempty"`
-	// If specified, the pod's scheduling constraints.
-	Affinity *v1.Affinity `json:"affinity,omitempty"`
-	// If specified, the pod's tolerations.
-	Tolerations []v1.Toleration `json:"tolerations,omitempty"`
-	// If specified, the pod's topology spread constraints.
-	TopologySpreadConstraints []v1.TopologySpreadConstraint `json:"topologySpreadConstraints,omitempty"`
-	// SecurityContext holds pod-level security attributes and common container settings.
-	// This defaults to the default PodSecurityContext.
-	SecurityContext *v1.PodSecurityContext `json:"securityContext,omitempty"`
-	// ServiceAccountName is the name of the ServiceAccount to use to run the
-	// Prometheus Pods.
-	ServiceAccountName string `json:"serviceAccountName,omitempty"`
-	// ListenLocal makes the Alertmanager server listen on loopback, so that it
-	// does not bind against the Pod IP. Note this is only for the Alertmanager
-	// UI, not the gossip communication.
-	ListenLocal bool `json:"listenLocal,omitempty"`
-	// Containers allows injecting additional containers. This is meant to
-	// allow adding an authentication proxy to an Alertmanager pod.
-	// Containers described here modify an operator generated container if they
-	// share the same name and modifications are done via a strategic merge
-	// patch. The current container names are: `alertmanager` and
-	// `config-reloader`. Overriding containers is entirely outside the scope
-	// of what the maintainers will support and by doing so, you accept that
-	// this behaviour may break at any time without notice.
-	Containers []v1.Container `json:"containers,omitempty"`
-	// InitContainers allows adding initContainers to the pod definition. Those can be used to e.g.
-	// fetch secrets for injection into the Alertmanager configuration from external sources. Any
-	// errors during the execution of an initContainer will lead to a restart of the Pod. More info: https://kubernetes.io/docs/concepts/workloads/pods/init-containers/
-	// Using initContainers for any use case other then secret fetching is entirely outside the scope
-	// of what the maintainers will support and by doing so, you accept that this behaviour may break
-	// at any time without notice.
-	InitContainers []v1.Container `json:"initContainers,omitempty"`
-	// Priority class assigned to the Pods
-	PriorityClassName string `json:"priorityClassName,omitempty"`
-	// AdditionalPeers allows injecting a set of additional Alertmanagers to peer with to form a highly available cluster.
-	AdditionalPeers []string `json:"additionalPeers,omitempty"`
-	// ClusterAdvertiseAddress is the explicit address to advertise in cluster.
-	// Needs to be provided for non RFC1918 [1] (public) addresses.
-	// [1] RFC1918: https://tools.ietf.org/html/rfc1918
-	ClusterAdvertiseAddress string `json:"clusterAdvertiseAddress,omitempty"`
-	// Interval between gossip attempts.
-	ClusterGossipInterval GoDuration `json:"clusterGossipInterval,omitempty"`
-	// Interval between pushpull attempts.
-	ClusterPushpullInterval GoDuration `json:"clusterPushpullInterval,omitempty"`
-	// Timeout for cluster peering.
-	ClusterPeerTimeout GoDuration `json:"clusterPeerTimeout,omitempty"`
-	// Port name used for the pods and governing service.
-	// This defaults to web
-	PortName string `json:"portName,omitempty"`
-	// ForceEnableClusterMode ensures Alertmanager does not deactivate the cluster mode when running with a single replica.
-	// Use case is e.g. spanning an Alertmanager cluster across Kubernetes clusters with a single replica in each.
-	ForceEnableClusterMode bool `json:"forceEnableClusterMode,omitempty"`
-	// AlertmanagerConfigs to be selected for to merge and configure Alertmanager with.
-	AlertmanagerConfigSelector *metav1.LabelSelector `json:"alertmanagerConfigSelector,omitempty"`
-	// The AlertmanagerConfigMatcherStrategy defines how AlertmanagerConfig objects match the alerts.
-	// In the future more options may be added.
-	AlertmanagerConfigMatcherStrategy AlertmanagerConfigMatcherStrategy `json:"alertmanagerConfigMatcherStrategy,omitempty"`
-	// Namespaces to be selected for AlertmanagerConfig discovery. If nil, only
-	// check own namespace.
-	AlertmanagerConfigNamespaceSelector *metav1.LabelSelector `json:"alertmanagerConfigNamespaceSelector,omitempty"`
-	// Minimum number of seconds for which a newly created pod should be ready
-	// without any of its container crashing for it to be considered available.
-	// Defaults to 0 (pod will be considered available as soon as it is ready)
-	// This is an alpha field and requires enabling StatefulSetMinReadySeconds feature gate.
-	// +optional
-	MinReadySeconds *uint32 `json:"minReadySeconds,omitempty"`
-	// Pods' hostAliases configuration
-	// +listType=map
-	// +listMapKey=ip
-	HostAliases []HostAlias `json:"hostAliases,omitempty"`
-	// Defines the web command line flags when starting Alertmanager.
-	Web *AlertmanagerWebSpec `json:"web,omitempty"`
-	// EXPERIMENTAL: alertmanagerConfiguration specifies the configuration of Alertmanager.
-	// If defined, it takes precedence over the `configSecret` field.
-	// This field may change in future releases.
-	AlertmanagerConfiguration *AlertmanagerConfiguration `json:"alertmanagerConfiguration,omitempty"`
-}
-
-// AlertmanagerConfigMatcherStrategy defines the strategy used by AlertmanagerConfig objects to match alerts.
-type AlertmanagerConfigMatcherStrategy struct {
-	// If set to `OnNamespace`, the operator injects a label matcher matching the namespace of the AlertmanagerConfig object for all its routes and inhibition rules.
-	// `None` will not add any additional matchers other than the ones specified in the AlertmanagerConfig.
-	// Default is `OnNamespace`.
-	// +kubebuilder:validation:Enum="OnNamespace";"None"
-	// +kubebuilder:default:="OnNamespace"
-	Type string `json:"type,omitempty"`
-}
-
-// AlertmanagerConfiguration defines the Alertmanager configuration.
-type AlertmanagerConfiguration struct {
-	// The name of the AlertmanagerConfig resource which is used to generate the Alertmanager configuration.
-	// It must be defined in the same namespace as the Alertmanager object.
-	// The operator will not enforce a `namespace` label for routes and inhibition rules.
-	// +kubebuilder:validation:MinLength=1
-	Name string `json:"name,omitempty"`
-	// Defines the global parameters of the Alertmanager configuration.
-	// +optional
-	Global *AlertmanagerGlobalConfig `json:"global,omitempty"`
-	// Custom notification templates.
-	// +optional
-	Templates []SecretOrConfigMap `json:"templates,omitempty"`
-}
-
-// AlertmanagerGlobalConfig configures parameters that are valid in all other configuration contexts.
-// See https://prometheus.io/docs/alerting/latest/configuration/#configuration-file
-type AlertmanagerGlobalConfig struct {
-	// ResolveTimeout is the default value used by alertmanager if the alert does
-	// not include EndsAt, after this time passes it can declare the alert as resolved if it has not been updated.
-	// This has no impact on alerts from Prometheus, as they always include EndsAt.
-	ResolveTimeout Duration `json:"resolveTimeout,omitempty"`
-
-	// HTTP client configuration.
-	HTTPConfig *HTTPConfig `json:"httpConfig,omitempty"`
-}
-
-// HTTPConfig defines a client HTTP configuration.
-// See https://prometheus.io/docs/alerting/latest/configuration/#http_config
-type HTTPConfig struct {
-	// Authorization header configuration for the client.
-	// This is mutually exclusive with BasicAuth and is only available starting from Alertmanager v0.22+.
-	// +optional
-	Authorization *SafeAuthorization `json:"authorization,omitempty"`
-	// BasicAuth for the client.
-	// This is mutually exclusive with Authorization. If both are defined, BasicAuth takes precedence.
-	// +optional
-	BasicAuth *BasicAuth `json:"basicAuth,omitempty"`
-	// OAuth2 client credentials used to fetch a token for the targets.
-	// +optional
-	OAuth2 *OAuth2 `json:"oauth2,omitempty"`
-	// The secret's key that contains the bearer token to be used by the client
-	// for authentication.
-	// The secret needs to be in the same namespace as the Alertmanager
-	// object and accessible by the Prometheus Operator.
-	// +optional
-	BearerTokenSecret *v1.SecretKeySelector `json:"bearerTokenSecret,omitempty"`
-	// TLS configuration for the client.
-	// +optional
-	TLSConfig *SafeTLSConfig `json:"tlsConfig,omitempty"`
-	// Optional proxy URL.
-	// +optional
-	ProxyURL string `json:"proxyURL,omitempty"`
-	// FollowRedirects specifies whether the client should follow HTTP 3xx redirects.
-	// +optional
-	FollowRedirects *bool `json:"followRedirects,omitempty"`
-}
-
-// +kubebuilder:object:root=true
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
-// AlertmanagerList is a list of Alertmanagers.
-type AlertmanagerList struct {
-	metav1.TypeMeta `json:",inline"`
-	// Standard list metadata
-	// More info: https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#metadata
-	metav1.ListMeta `json:"metadata,omitempty"`
-	// List of Alertmanagers
-	Items []Alertmanager `json:"items"`
-}
-
-// MetadataConfig configures the sending of series metadata to the remote storage.
-type MetadataConfig struct {
-	// Whether metric metadata is sent to the remote storage or not.
-	Send bool `json:"send,omitempty"`
-	// How frequently metric metadata is sent to the remote storage.
-	SendInterval Duration `json:"sendInterval,omitempty"`
-}
-
-// AlertmanagerStatus is the most recent observed status of the Alertmanager cluster. Read-only. Not
-// included when requesting from the apiserver, only from the Prometheus
-// Operator API itself. More info:
-// https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
-type AlertmanagerStatus struct {
-	// Represents whether any actions on the underlying managed objects are
-	// being performed. Only delete actions will be performed.
-	Paused bool `json:"paused"`
-	// Total number of non-terminated pods targeted by this Alertmanager
-	// cluster (their labels match the selector).
-	Replicas int32 `json:"replicas"`
-	// Total number of non-terminated pods targeted by this Alertmanager
-	// cluster that have the desired version spec.
-	UpdatedReplicas int32 `json:"updatedReplicas"`
-	// Total number of available pods (ready for at least minReadySeconds)
-	// targeted by this Alertmanager cluster.
-	AvailableReplicas int32 `json:"availableReplicas"`
-	// Total number of unavailable pods targeted by this Alertmanager cluster.
-	UnavailableReplicas int32 `json:"unavailableReplicas"`
 }
 
 // NamespaceSelector is a selector for selecting either all namespaces or a
@@ -1955,45 +676,6 @@ type NamespaceSelector struct {
 	// implementation to support label selections.
 }
 
-// /--rules.*/ command-line arguments
-type Rules struct {
-	Alert RulesAlert `json:"alert,omitempty"`
-}
-
-// /--rules.alert.*/ command-line arguments
-type RulesAlert struct {
-	// Max time to tolerate prometheus outage for restoring 'for' state of alert.
-	ForOutageTolerance string `json:"forOutageTolerance,omitempty"`
-	// Minimum duration between alert and restored 'for' state.
-	// This is maintained only for alerts with configured 'for' time greater than grace period.
-	ForGracePeriod string `json:"forGracePeriod,omitempty"`
-	// Minimum amount of time to wait before resending an alert to Alertmanager.
-	ResendDelay string `json:"resendDelay,omitempty"`
-}
-
-// ProbeTLSConfig specifies TLS configuration parameters for the prober.
-type ProbeTLSConfig struct {
-	SafeTLSConfig `json:",inline"`
-}
-
-// SafeAuthorization specifies a subset of the Authorization struct, that is
-// safe for use in Endpoints (no CredentialsFile field)
-type SafeAuthorization struct {
-	// Set the authentication type. Defaults to Bearer, Basic will cause an
-	// error
-	Type string `json:"type,omitempty"`
-	// The secret's key that contains the credentials of the request
-	Credentials *v1.SecretKeySelector `json:"credentials,omitempty"`
-}
-
-// Authorization contains optional `Authorization` header configuration.
-// This section is only understood by versions of Prometheus >= 2.26.0.
-type Authorization struct {
-	SafeAuthorization `json:",inline"`
-	// File to read a secret from, mutually exclusive with Credentials (from SafeAuthorization)
-	CredentialsFile string `json:"credentialsFile,omitempty"`
-}
-
 // Argument as part of the AdditionalArgs list.
 type Argument struct {
 	// Name of the argument, e.g. "scrape.discovery-reload-interval".
@@ -2002,3 +684,44 @@ type Argument struct {
 	// Argument value, e.g. 30s. Can be empty for name-only arguments (e.g. --storage.tsdb.no-lockfile)
 	Value string `json:"value,omitempty"`
 }
+
+// The valid options for Role.
+const (
+	RoleNode          = "node"
+	RolePod           = "pod"
+	RoleService       = "service"
+	RoleEndpoint      = "endpoints"
+	RoleEndpointSlice = "endpointslice"
+	RoleIngress       = "ingress"
+)
+
+// NativeHistogramConfig extends the native histogram configuration settings.
+type NativeHistogramConfig struct {
+	// Whether to scrape a classic histogram that is also exposed as a native histogram.
+	// It requires Prometheus >= v2.45.0.
+	//
+	// +optional
+	ScrapeClassicHistograms *bool `json:"scrapeClassicHistograms,omitempty"`
+
+	// If there are more than this many buckets in a native histogram,
+	// buckets will be merged to stay within the limit.
+	// It requires Prometheus >= v2.45.0.
+	//
+	// +optional
+	NativeHistogramBucketLimit *uint64 `json:"nativeHistogramBucketLimit,omitempty"`
+
+	// If the growth factor of one bucket to the next is smaller than this,
+	// buckets will be merged to increase the factor sufficiently.
+	// It requires Prometheus >= v2.50.0.
+	//
+	// +optional
+	NativeHistogramMinBucketFactor *resource.Quantity `json:"nativeHistogramMinBucketFactor,omitempty"`
+}
+
+// +kubebuilder:validation:Enum=RelabelConfig;RoleSelector
+type SelectorMechanism string
+
+const (
+	SelectorMechanismRelabel SelectorMechanism = "RelabelConfig"
+	SelectorMechanismRole    SelectorMechanism = "RoleSelector"
+)
